Test Dataset California Test Score Data

420 school level observations a cross-section from 1998-1999 distcod, district code county, county district, district grspan, grade span of district enrltot, total enrollment teachers, number of teachers calwpct, percent qualifying for CalWorks mealpct, percent qualifying for reduced-price lunch computer, number of computers testscr, average test score (read.scr+math.scr)/2 compstu, computer per student expnstu, expenditure per student str, student teacher ratio avginc, district average income elpct, percent of English learners readscr, average reading score mathscr, average math score

data(df_opt_caschool)

Format

csv

Source

https://fmwww.bc.edu/ec-p/data/stockwatson/caschool.des

References

Stock, James H. and Mark W. Watson (2003) Introduction to Econometrics, Addison-Wesley Educational Publishers, chapter 4–7.

Examples

data(df_opt_caschool)
head(df_opt_caschool, 10)
#>     X distcod      county                        district grspan enrltot
#> 1   1   75119     Alameda              Sunol Glen Unified  KK-08     195
#> 2   2   61499       Butte            Manzanita Elementary  KK-08     240
#> 3   3   61549       Butte     Thermalito Union Elementary  KK-08    1550
#> 4   4   61457       Butte Golden Feather Union Elementary  KK-08     243
#> 5   5   61523       Butte        Palermo Union Elementary  KK-08    1335
#> 6   6   62042      Fresno         Burrel Union Elementary  KK-08     137
#> 7   7   68536 San Joaquin           Holt Union Elementary  KK-08     195
#> 8   8   63834        Kern             Vineland Elementary  KK-08     888
#> 9   9   62331      Fresno        Orange Center Elementary  KK-08     379
#> 10 10   67306  Sacramento     Del Paso Heights Elementary  KK-06    2247
#>    teachers calwpct  mealpct computer testscr    compstu  expnstu      str
#> 1     10.90  0.5102   2.0408       67  690.80 0.34358975 6384.911 17.88991
#> 2     11.15 15.4167  47.9167      101  661.20 0.42083332 5099.381 21.52466
#> 3     82.90 55.0323  76.3226      169  643.60 0.10903226 5501.955 18.69723
#> 4     14.00 36.4754  77.0492       85  647.70 0.34979424 7101.831 17.35714
#> 5     71.50 33.1086  78.4270      171  640.85 0.12808989 5235.988 18.67133
#> 6      6.40 12.3188  86.9565       25  605.55 0.18248175 5580.147 21.40625
#> 7     10.00 12.9032  94.6237       28  606.75 0.14358975 5253.331 19.50000
#> 8     42.50 18.8063 100.0000       66  609.00 0.07432432 4565.746 20.89412
#> 9     19.00 32.1900  93.1398       35  612.50 0.09234829 5355.548 19.94737
#> 10   108.00 78.9942  87.3164        0  612.65 0.00000000 5036.211 20.80556
#>       avginc     elpct readscr mathscr
#> 1  22.690001  0.000000   691.6   690.0
#> 2   9.824000  4.583333   660.5   661.9
#> 3   8.978000 30.000002   636.3   650.9
#> 4   8.978000  0.000000   651.9   643.5
#> 5   9.080333 13.857677   641.8   639.9
#> 6  10.415000 12.408759   605.7   605.4
#> 7   6.577000 68.717949   604.5   609.0
#> 8   8.174000 46.959461   605.5   612.5
#> 9   7.385000 30.079157   608.9   616.1
#> 10 11.613333 40.275921   611.9   613.4