Test Dataset California Test Score Data: IB Frame, Alternative Allocation

\(\text{ID}_i\): individual id \(A_{i,l=0}\): A when considering the lth allocation for ith person, expected outcome without the lth allocation \(\alpha^{\text{o}}_{i}\): see df_casch_prep_i, observed/random/uniform allocation's marginal effect. Note since \(\alpha_{il}\) considers only additional allocation, with existing allocations embeded in \(D_{il}\), \(\alpha^{\text{o}}_{i}\) also only includes additional allocations that is not already embeded in \(\A_{i,l=0}\). \(\beta_{i}\): i specific preference

data(df_opt_caschool_input_ib)

Format

csv

Source

https://fmwww.bc.edu/ec-p/data/stockwatson/caschool.des https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_sodis_rkone_casch_allrw.html

References

Stock, James H. and Mark W. Watson (2003) Introduction to Econometrics, Addison-Wesley Educational Publishers, chapter 4–7.

Examples

data(df_opt_caschool_input_ib)
head(df_opt_caschool_input_ib, 30)

#>    id_i   A_i_l0 alpha_o_i      beta_i
#> 1     1 690.8000  2.372592 0.002380952
#> 2     2 654.2762  2.920114 0.002380952
#> 3     3 640.8497  2.761831 0.002380952
#> 4     4 641.2885  2.664680 0.002380952
#> 5     5 639.3850  2.625868 0.002380952
#> 6     6 625.4803  2.837696 0.002380952
#> 7     7 612.8571  2.827339 0.002380952
#> 8     8 617.2157  2.942918 0.002380952
#> 9     9 624.6470  3.017952 0.002380952
#> 10   10 621.9968  3.063044 0.002380952
#> 11   11 626.0318  2.956175 0.002380952
#> 12   12 621.1129  2.936083 0.002380952
#> 13   13 621.7829  2.986830 0.002380952
#> 14   14 634.4715  3.043577 0.002380952
#> 15   15 636.8020  2.553091 0.002380952
#> 16   16 624.3424  2.823611 0.002380952
#> 17   17 622.0817  2.524999 0.002380952
#> 18   18 634.4041  2.272017 0.002380952
#> 19   19 620.7221  3.310311 0.002380952
#> 20   20 622.3484  2.886947 0.002380952
#> 21   21 622.0775  2.749844 0.002380952
#> 22   22 623.8981  3.148392 0.002380952
#> 23   23 635.8420  2.830685 0.002380952
#> 24   24 617.1388  3.790356 0.002380952
#> 25   25 623.6784  3.019650 0.002380952
#> 26   26 631.0983  2.708787 0.002380952
#> 27   27 617.4778  3.310343 0.002380952
#> 28   28 623.3784  2.793508 0.002380952
#> 29   29 627.5394  2.843753 0.002380952
#> 30   30 631.8025  2.749774 0.002380952