Test Dataset California Test Score Data: PREP I Frame, Select Input and Estimation Dates

\(\text{ID}_i\): individual id \(D^{\text{max}}_{i}\): maximimum discrete allocation each person, in addition to existing levels \(D^{\text{o}}_{i}\): what I call observed allocation, just any comparison allocation, in addition to existing levels, this is the additional allocation (ignoring existing levels), this is the not the total allocation. If fully redistributing, this could be the observed level of allocation. \(\Omega_{i}\): This is the expected outcome when the input of interest is completely zero, \(\Omega_{i}\neq A_{i,l=0}\), because \(l=0\) does not mean input is zero, using \(\Omega_i\) rather than \(A_i\) to avoid confusion because \(A_i\) sounds like \(A_i = A_{i, l=0}\), but again, the problem does not start with allocation at zero, but each district already has some existing number of teachers. \(\theta_{i}\): coefficient in front of student teacher ratio \(\beta_{i}\): i specific preference \(\text{enrltot}_{i}\): enrollment per district/school, assume no within district variations \(\text{teachers}_{i}\): teacher per distrct/school \(\text{stravg}_{i}\): average of student teacher ratio district/school

data(df_opt_caschool_prep_i)

Format

csv

Source

https://fmwww.bc.edu/ec-p/data/stockwatson/caschool.des https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_sodis_rkone_casch_allrw.html

References

Stock, James H. and Mark W. Watson (2003) Introduction to Econometrics, Addison-Wesley Educational Publishers, chapter 4–7.

Examples

data(df_opt_caschool_prep_i)
head(df_opt_caschool_prep_i, 30)
#>    id_i D_max_i D_o_i  Omega_i    theta_i      beta_i enrltot teachers   stravg
#> 1     1      11     2 706.2218 -0.8699505 0.002380952     195       11 17.72727
#> 2     2      11     2 673.2570 -0.8699505 0.002380952     240       11 21.81818
#> 3     3      83    17 657.0958 -0.8699505 0.002380952    1550       83 18.67470
#> 4     4      14     3 656.3883 -0.8699505 0.002380952     243       14 17.35714
#> 5     5      72    14 655.5153 -0.8699505 0.002380952    1335       72 18.54167
#> 6     6       6     1 645.3442 -0.8699505 0.002380952     137        6 22.83333
#> 7     7      10     2 629.8211 -0.8699505 0.002380952     195       10 19.50000
#> 8     8      42     8 635.6090 -0.8699505 0.002380952     888       42 21.14286
#> 9     9      19     4 642.0002 -0.8699505 0.002380952     379       19 19.94737
#> 10   10     108    22 640.0966 -0.8699505 0.002380952    2247      108 20.80556
#> 11   11      21     4 644.5079 -0.8699505 0.002380952     446       21 21.23810
#> 12   12      47     9 639.3819 -0.8699505 0.002380952     987       47 21.00000
#> 13   13       5     1 639.7038 -0.8699505 0.002380952     103        5 20.60000
#> 14   14      24     5 652.1242 -0.8699505 0.002380952     487       24 20.29167
#> 15   15      36     7 652.4853 -0.8699505 0.002380952     649       36 18.02778
#> 16   16      42     8 641.9899 -0.8699505 0.002380952     852       42 20.28571
#> 17   17      29     6 636.8108 -0.8699505 0.002380952     491       29 16.93103
#> 18   18      26     5 648.4906 -0.8699505 0.002380952     421       26 16.19231
#> 19   19     303    61 640.4754 -0.8699505 0.002380952    6880      303 22.70627
#> 20   20     135    27 639.6701 -0.8699505 0.002380952    2688      135 19.91111
#> 21   21      24     5 638.0266 -0.8699505 0.002380952     440       24 18.33333
#> 22   22      21     4 643.5756 -0.8699505 0.002380952     475       21 22.61905
#> 23   23     130    26 652.8261 -0.8699505 0.002380952    2538      130 19.52308
#> 24   24      19     4 638.9334 -0.8699505 0.002380952     476       19 25.05263
#> 25   25     114    23 641.6650 -0.8699505 0.002380952    2357      114 20.67544
#> 26   26      85    17 647.3510 -0.8699505 0.002380952    1588       85 18.68235
#> 27   27     320    64 637.3398 -0.8699505 0.002380952    7306      320 22.83125
#> 28   28     135    27 640.1394 -0.8699505 0.002380952    2601      135 19.26667
#> 29   29      44     9 644.2859 -0.8699505 0.002380952     847       44 19.25000
#> 30   30      22     4 649.6760 -0.8699505 0.002380952     452       22 20.54545
#>    distcod      county                        district grspan    avginc testscr
#> 1    75119     Alameda              Sunol Glen Unified  KK-08 22.690001  690.80
#> 2    61499       Butte            Manzanita Elementary  KK-08  9.824000  661.20
#> 3    61549       Butte     Thermalito Union Elementary  KK-08  8.978000  643.60
#> 4    61457       Butte Golden Feather Union Elementary  KK-08  8.978000  647.70
#> 5    61523       Butte        Palermo Union Elementary  KK-08  9.080333  640.85
#> 6    62042      Fresno         Burrel Union Elementary  KK-08 10.415000  605.55
#> 7    68536 San Joaquin           Holt Union Elementary  KK-08  6.577000  606.75
#> 8    63834        Kern             Vineland Elementary  KK-08  8.174000  609.00
#> 9    62331      Fresno        Orange Center Elementary  KK-08  7.385000  612.50
#> 10   67306  Sacramento     Del Paso Heights Elementary  KK-06 11.613333  612.65
#> 11   65722      Merced       Le Grand Union Elementary  KK-08  8.931000  615.75
#> 12   62174      Fresno          West Fresno Elementary  KK-08  7.385000  616.30
#> 13   71795      Tulare          Allensworth Elementary  KK-08  5.335000  616.30
#> 14   72181      Tulare      Sunnyside Union Elementary  KK-08  8.279000  616.30
#> 15   72298      Tulare            Woodville Elementary  KK-08  9.630000  616.45
#> 16   72041      Tulare         Pixley Union Elementary  KK-08  7.454000  617.35
#> 17   63594        Kern     Lost Hills Union Elementary  KK-08  6.216000  618.05
#> 18   63370        Kern   Buttonwillow Union Elementary  KK-08  7.764000  618.30
#> 19   64709 Los Angeles               Lennox Elementary  KK-08  7.022000  619.80
#> 20   63560        Kern               Lamont Elementary  KK-08  5.699000  620.30
#> 21   63230    Imperial    Westmorland Union Elementary  KK-08  7.941000  620.50
#> 22   72058      Tulare        Pleasant View Elementary  KK-08  9.630000  621.40
#> 23   63842        Kern          Wasco Union Elementary  KK-08  7.405000  621.75
#> 24   71811      Tulare           Alta Vista Elementary  KK-08  9.630000  622.05
#> 25   65748      Merced     Livingston Union Elementary  KK-08  8.019000  622.60
#> 26   72272      Tulare       Woodlake Union Elementary  KK-08  8.523000  623.10
#> 27   65961    Monterey         Alisal Union Elementary  KK-06  7.983181  623.20
#> 28   63313        Kern          Arvin Union Elementary  KK-08  7.305000  623.45
#> 29   72199      Tulare    Terra Bella Union Elementary  KK-08  8.934000  623.60
#> 30   72215      Tulare               Tipton Elementary  KK-08  8.554000  624.15