Test Dataset Job Training Effects Lalonde AER 1986

trt, a numeric vector identifying the study in which the subjects were enrolled (0 = Control, 1 = treated). Male module, 297 treated. 425 untreated. age, age (in years). educ, years of education. black, (0 = not black, 1 = black). hisp, (0 = not hispanic, 1 = hispanic). marr, (0 = not married, 1 = married). nodeg, (0 = completed high school, 1 = dropout). re74, real earnings in 1974. re75, real earnings in 1975. re78, real earnings in 1978.

data(df_opt_lalonde_training)

Format

csv

Source

https://www.jstor.org/stable/1806062?

References

Lalonde, R. (1986) "Evaluating the Econometric Evaluations of Training Programs with Experimental Data", American Economic Review, 604–620.

Examples

library(REconTools)
data(df_opt_lalonde_training)
head(df_opt_lalonde_training, 10)
#>     X trt age educ black hisp marr nodeg re74     re75      re78
#> 1   1   0  23   10     1    0    0     1    0    0.000     0.000
#> 2   2   0  26   12     0    0    0     0    0    0.000 12383.680
#> 3   3   0  22    9     1    0    0     1    0    0.000     0.000
#> 4   4   0  34    9     1    0    0     1   NA 4368.413 14051.160
#> 5   5   0  18    9     1    0    0     1    0    0.000 10740.080
#> 6   6   0  45   11     1    0    0     1    0    0.000 11796.470
#> 7   7   0  18    9     1    0    0     1    0    0.000  9227.052
#> 8   8   0  27   12     1    0    0     0   NA 2483.851  3552.268
#> 9   9   0  24    8     0    0    0     1    0    0.000 10569.270
#> 10 10   0  34   11     1    0    1     1    0    0.000  6040.335
REconTools::ff_summ_percentiles(df_opt_lalonde_training)
#> Warning: `funs()` was deprecated in dplyr 0.8.0.
#> Please use a list of either functions or lambdas: 
#> 
#>   # Simple named list: 
#>   list(mean = mean, median = median)
#> 
#>   # Auto named with `tibble::lst()`: 
#>   tibble::lst(mean, median)
#> 
#>   # Using lambdas
#>   list(~ mean(., trim = .2), ~ median(., na.rm = TRUE))
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.
#> Warning: `as.tibble()` was deprecated in tibble 2.0.0.
#> Please use `as_tibble()` instead.
#> The signature and semantics have changed, see `?as_tibble`.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.
#> Warning: attributes are not identical across measure variables;
#> they will be dropped
#> Warning: The `x` argument of `as_tibble.matrix()` must have unique column names if `.name_repair` is omitted as of tibble 2.0.0.
#> Using compatibility `.name_repair`.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_lifecycle_warnings()` to see where this warning was generated.
#> # A tibble: 18 x 12
#>    stats   X      age    black  educ   hisp  marr  nodeg re74  re75  re78  trt  
#>    <chr>   <chr>  <chr>  <chr>  <chr>  <chr> <chr> <chr> <chr> <chr> <chr> <chr>
#>  1 n       "722"  "722"  "722"  "722"  "722" "722" "722" "722" "722" "722" "722"
#>  2 unique  "722"  " 35"  "  2"  " 14"  "  2" "  2" "  2" "116" "424" "524" "  2"
#>  3 NAobs   "  0"  "  0"  "  0"  "  0"  "  0" "  0" "  0" "277" "  0" "  0" "  0"
#>  4 ZEROobs "  0"  "  0"  "144"  "  0"  "646" "605" "159"  NA   "289" "196" "425"
#>  5 mean    " 854~ "  24~ "   0~ "  10~ "   ~ "   ~ "   ~ "210~ "304~ "545~ "   ~
#>  6 sd      " 934~ "   6~ "   0~ "   1~ "   ~ "   ~ "   ~ "536~ "506~ "625~ "   ~
#>  7 cv      "1.09~ "0.27~ "0.49~ "0.16~ "2.9~ "2.2~ "0.5~ "2.5~ "1.6~ "1.1~ "1.1~
#>  8 min     " 1"   "17"   " 0"   " 3"   " 0"  " 0"  " 0"  " 0"  " 0"  " 0"  " 0" 
#>  9 p01     " 8.2~ "17.0~ " 0.0~ " 5.0~ " 0.~ " 0.~ " 0.~ " 0.~ " 0.~ " 0.~ " 0.~
#> 10 p05     "37.0~ "17.0~ " 0.0~ " 8.0~ " 0.~ " 0.~ " 0.~ " 0.~ " 0.~ " 0.~ " 0.~
#> 11 p10     "73.1" "18.0" " 0.0" " 8.0" " 0.~ " 0.~ " 0.~ " 0.~ " 0.~ " 0.~ " 0.~
#> 12 p25     "181.~ " 19.~ "  1.~ "  9.~ "  0~ "  0~ "  1~ "  0~ "  0~ "  0~ "  0~
#> 13 p50     " 361~ "  23~ "   1~ "  10~ "   ~ "   ~ "   ~ "   ~ " 93~ "395~ "   ~
#> 14 p75     "1458~ "  27~ "   1~ "  11~ "   ~ "   ~ "   ~ " 82~ "399~ "877~ "   ~
#> 15 p90     " 243~ "   3~ "    ~ "   1~ "   ~ "   ~ "   ~ " 78~ " 89~ "128~ "   ~
#> 16 p95     " 276~ "   3~ "    ~ "   1~ "   ~ "   ~ "   ~ "122~ "122~ "166~ "   ~
#> 17 p99     " 338~ "   4~ "    ~ "   1~ "   ~ "   ~ "   ~ "270~ "242~ "260~ "   ~
#> 18 max     " 411~ "   5~ "    ~ "   1~ "   ~ "   ~ "   ~ "395~ "374~ "603~ "   ~