Lalonde AER 1986 with Wage Regression Allocation Queue Queue

appended onto the raw Lalonde binary targeting queue rank info based on different planner inequality aversion. includes min and max rank. Based on training logit regressin.

data(df_opt_lalonde_training_employ)

Format

csv

Source

https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_sobin_rkone_allfc_training_logit.html

References

Lalonde, R. (1986) "Evaluating the Econometric Evaluations of Training Programs with Experimental Data", American Economic Review, 604–620.

Examples

data(df_opt_lalonde_training_employ)
head(df_opt_lalonde_training_employ, 10)

#>    id A_i_employ alpha_i_employ beta_i_employ rank_min_employ rank_max_employ
#> 1   1  0.6466332     0.09434268   0.001385042             194               8
#> 2   2  0.8485082     0.03809367   0.001385042             682             620
#> 3   3  0.6499830     0.09380271   0.001385042             199              13
#> 4   4  0.5651946     0.07950350   0.001385042             260               3
#> 5   5  0.6824193     0.08817794   0.001385042             383             130
#> 6   6  0.6433307     0.07240217   0.001385042             462             185
#> 7   7  0.6824193     0.08817794   0.001385042             383             130
#> 8   8  0.6295665     0.07390839   0.001385042             434             158
#> 9   9  0.8260749     0.04286336   0.001385042             646             597
#> 10 10  0.6694840     0.06924924   0.001385042             503             294
#>    avg_rank_employ   trt age educ black hisp marr nodeg re74     re75      re78
#> 1          81.9000 ntran  23   10     1    0    0     1    0    0.000     0.000
#> 2         662.4667 ntran  26   12     0    0    0     0    0    0.000 12383.680
#> 3          88.6000 ntran  22    9     1    0    0     1    0    0.000     0.000
#> 4         113.0333 ntran  34    9     1    0    0     1   NA 4368.413 14051.160
#> 5         241.0000 ntran  18    9     1    0    0     1    0    0.000 10740.080
#> 6         352.6333 ntran  45   11     1    0    0     1    0    0.000 11796.470
#> 7         241.0000 ntran  18    9     1    0    0     1    0    0.000  9227.052
#> 8         312.1333 ntran  27   12     1    0    0     0   NA 2483.851  3552.268
#> 9         624.5667 ntran  24    8     0    0    0     1    0    0.000 10569.270
#> 10        433.5000 ntran  34   11     1    0    1     1    0    0.000  6040.335
#>    emp78 emp75 race age_m2 age_m3     p_mpg  p_mpg_hp p_mpg_hp_bi0 p_mpg_hp_bi1
#> 1      0     0    1      1      2 0.6783821 0.6466332    0.6466332    0.7409759
#> 2      1     0    0      2      2 0.8897568 0.8485082    0.8485082    0.8866018
#> 3      0     0    1      1      2 0.6883859 0.6499830    0.6499830    0.7437857
#> 4      1     1    1      2      3 0.6383402 0.5651946    0.5651946    0.6446980
#> 5      1     0    1      1      1 0.7042022 0.6824193    0.6824193    0.7705972
#> 6      1     0    1      2      3 0.5762456 0.6433307    0.6433307    0.7157329
#> 7      1     0    1      1      1 0.7042022 0.6824193    0.6824193    0.7705972
#> 8      1     1    1      2      3 0.7123512 0.6295665    0.6295665    0.7034749
#> 9      1     0    0      2      2 0.8749490 0.8260749    0.8260749    0.8689382
#> 10     1     0    1      2      3 0.6934077 0.6694840    0.6694840    0.7387332
#>    rho_c1_rk_employ rho_c2_rk_employ rho_c3_rk_employ rho_c4_rk_employ
#> 1                 8                8                8                8
#> 2               682              682              682              682
#> 3                13               13               13               13
#> 4               260              260              260              260
#> 5               130              130              130              130
#> 6               462              462              462              462
#> 7               130              130              130              130
#> 8               434              434              434              434
#> 9               646              646              646              646
#> 10              503              503              503              503
#>    rho_c5_rk_employ rho_c6_rk_employ rho_c7_rk_employ rho_c8_rk_employ
#> 1                 8                8                8                8
#> 2               682              682              682              682
#> 3                13               13               13               13
#> 4               256              256              256              256
#> 5               130              130              130              130
#> 6               462              462              462              462
#> 7               130              130              130              130
#> 8               434              430              430              430
#> 9               646              646              646              646
#> 10              503              503              503              501
#>    rho_c9_rk_employ rho_c10_rk_employ rho_c11_rk_employ rho_c12_rk_employ
#> 1                 8                 8                 8                 8
#> 2               681               681               681               680
#> 3                13                13                13                13
#> 4               245               243               233               226
#> 5               130               130               130               130
#> 6               462               461               457               455
#> 7               130               130               130               130
#> 8               430               428               428               425
#> 9               646               646               646               646
#> 10              501               501               501               500
#>    rho_c13_rk_employ rho_c14_rk_employ rho_c15_rk_employ rho_c16_rk_employ
#> 1                  8                 8                 8                 8
#> 2                679               679               677               675
#> 3                 13                13                13                16
#> 4                173               100                54                11
#> 5                130               148               219               255
#> 6                455               451               442               427
#> 7                130               148               219               255
#> 8                420               411               399               387
#> 9                644               641               636               632
#> 10               500               499               494               490
#>    rho_c17_rk_employ rho_c18_rk_employ rho_c19_rk_employ rho_c20_rk_employ
#> 1                 52               119               142               147
#> 2                673               670               667               664
#> 3                 81               126               148               156
#> 4                  3                 3                 3                 3
#> 5                269               289               316               317
#> 6                414               366               298               256
#> 7                269               289               316               317
#> 8                328               255               219               186
#> 9                624               616               613               604
#> 10               484               476               458               438
#>    rho_c21_rk_employ rho_c22_rk_employ rho_c23_rk_employ rho_c24_rk_employ
#> 1                162               173               185               185
#> 2                661               649               640               632
#> 3                172               181               190               199
#> 4                  3                 3                 3                 3
#> 5                331               349               371               378
#> 6                226               207               200               193
#> 7                331               349               371               378
#> 8                166               158               158               158
#> 9                600               599               597               597
#> 10               393               353               320               309
#>    rho_c25_rk_employ rho_c26_rk_employ rho_c27_rk_employ rho_c28_rk_employ
#> 1                194               194               194               194
#> 2                629               620               620               620
#> 3                199               199               199               199
#> 4                  3                 3                 3                 3
#> 5                383               383               383               383
#> 6                186               186               186               185
#> 7                383               383               383               383
#> 8                158               158               158               158
#> 9                597               597               597               597
#> 10               296               294               294               294
#>    rho_c29_rk_employ rho_c30_rk_employ
#> 1                194               194
#> 2                620               620
#> 3                199               199
#> 4                  3                 3
#> 5                383               383
#> 6                185               185
#> 7                383               383
#> 8                158               158
#> 9                597               597
#> 10               294               294