Lalonde AER 1986 with Wage Regression Allocation Queue Queue

appended onto the raw Lalonde binary targeting queue rank info based on different planner inequality aversion. includes min and max rank. Based on wage regression.

data(df_opt_lalonde_training_wage)

Format

csv

Source

https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_sobin_rkone_allrw_training_wage.html

References

Lalonde, R. (1986) "Evaluating the Econometric Evaluations of Training Programs with Experimental Data", American Economic Review, 604–620.

Examples

data(df_opt_lalonde_training_wage)
head(df_opt_lalonde_training_wage, 10)

#>    id A_i_wage alpha_i_wage beta_i_wage rank_min_wage rank_max_wage
#> 1   1 4.705623    0.4537963 0.001385042           284           392
#> 2   2 6.842885    0.4537963 0.001385042           673           689
#> 3   3 4.373738    0.4537963 0.001385042           163           278
#> 4   4 4.129823    0.4537963 0.001385042            61           178
#> 5   5 4.192084    0.4537963 0.001385042           106           223
#> 6   6 5.057681    0.4537963 0.001385042           399           483
#> 7   7 4.192084    0.4537963 0.001385042           106           223
#> 8   8 5.096506    0.4537963 0.001385042           416           499
#> 9   9 5.445291    0.4537963 0.001385042           495           545
#> 10 10 5.450645    1.0708127 0.001385042            68           496
#>    avg_rank_wage trt age educ race marr nodeg      re78 rho_c1_rk rho_c2_rk
#> 1       342.5667   0  23   10    1    0     1     0.000       284       284
#> 2       682.5333   0  26   12    0    0     0 12383.680       673       673
#> 3       220.7667   0  22    9    1    0     1     0.000       163       163
#> 4       116.7000   0  34    9    1    0     1 14051.160        61        61
#> 5       162.9000   0  18    9    1    0     1 10740.080       106       106
#> 6       445.5333   0  45   11    1    0     1 11796.470       399       399
#> 7       162.9000   0  18    9    1    0     1  9227.052       106       106
#> 8       462.3000   0  27   12    1    0     0  3552.268       416       416
#> 9       522.3333   0  24    8    0    0     1 10569.270       495       495
#> 10      270.3333   0  34   11    1    1     1  6040.335       496       496
#>    rho_c3_rk rho_c4_rk rho_c5_rk rho_c6_rk rho_c7_rk rho_c8_rk rho_c9_rk
#> 1        284       284       284       284       284       284       289
#> 2        673       673       673       673       673       673       673
#> 3        163       163       163       163       163       163       164
#> 4         61        61        61        61        61        61        61
#> 5        106       106       106       106       106       106       106
#> 6        399       399       399       399       399       399       400
#> 7        106       106       106       106       106       106       106
#> 8        416       416       416       416       416       416       418
#> 9        495       495       495       495       495       495       496
#> 10       496       496       496       496       496       496       494
#>    rho_c10_rk rho_c11_rk rho_c12_rk rho_c13_rk rho_c14_rk rho_c15_rk rho_c16_rk
#> 1         289        294        303        316        339        345        362
#> 2         674        674        675        684        688        689        689
#> 3         164        166        167        184        192        202        231
#> 4          61         61         61         68         69         90        115
#> 5         106        106        108        114        121        137        172
#> 6         401        403        415        436        450        453        460
#> 7         106        106        108        114        121        137        172
#> 8         419        423        436        453        466        469        477
#> 9         497        498        504        515        519        522        530
#> 10        491        485        481        380        351        309        199
#>    rho_c17_rk rho_c18_rk rho_c19_rk rho_c20_rk rho_c21_rk rho_c22_rk rho_c23_rk
#> 1         378        387        391        392        392        392        392
#> 2         689        689        689        689        689        689        689
#> 3         255        263        273        278        278        278        278
#> 4         150        156        164        177        178        178        178
#> 5         197        205        214        223        223        223        223
#> 6         478        482        483        483        483        483        483
#> 7         197        205        214        223        223        223        223
#> 8         494        498        499        499        499        499        499
#> 9         544        545        545        545        545        545        545
#> 10         68         68         68         68         68         68         68
#>    rho_c24_rk rho_c25_rk rho_c26_rk rho_c27_rk rho_c28_rk rho_c29_rk rho_c30_rk
#> 1         392        392        392        392        392        392        392
#> 2         689        689        689        689        689        689        689
#> 3         278        278        278        278        278        278        278
#> 4         178        178        178        178        178        178        178
#> 5         223        223        223        223        223        223        223
#> 6         483        483        483        483        483        483        483
#> 7         223        223        223        223        223        223        223
#> 8         499        499        499        499        499        499        499
#> 9         545        545        545        545        545        545        545
#> 10         68         68         68         68         68         68         68