Test Dataset North Carolina Crimes

A panel of 90 observations from 1981 to 1987 county, county identifier year, year from 1981 to 1987 crmrte, crimes committed per person prbarr, 'probability' of arrest prbconv, 'probability' of conviction prbpris, 'probability' of prison sentence avgsen, average sentence, days polpc, police per capita density, people per square mile taxpc, tax revenue per capita region, one of 'other', 'west' or 'central' smsa, 'yes' or 'no' if in SMSA pctmin, percentage minority in 1980 wcon, weekly wage in construction wtuc, weekly wage in trns, util, commun wtrd, weekly wage in whole sales and retail trade wfir, weekly wage in finance, insurance and real estate wser, weekly wage in service industry wmfg, weekly wage in manufacturing wfed, weekly wage of federal employees wsta, weekly wage of state employees wloc, weekly wage of local governments employees mix, offence mix: face-to-face/other pctymle, percentage of young males

data(df_opt_nccrimes)

Format

csv

Source

https://www.jstor.org/stable/2109893?

References

Cornwell, C. and W.N. Trumbull (1994) “Estimating the economic model of crime with panel data”, Review of Economics and Statistics, 76, 360–366.

Examples

data(df_opt_nccrimes)
head(df_opt_nccrimes, 10)
#>     X county year    crmrte   prbarr  prbconv  prbpris avgsen     polpc
#> 1   1      1   81 0.0398849 0.289696 0.402062 0.472222   5.61 0.0017868
#> 2   2      1   82 0.0383449 0.338111 0.433005 0.506993   5.59 0.0017666
#> 3   3      1   83 0.0303048 0.330449 0.525703 0.479705   5.80 0.0018358
#> 4   4      1   84 0.0347259 0.362525 0.604706 0.520104   6.89 0.0018859
#> 5   5      1   85 0.0365730 0.325395 0.578723 0.497059   6.55 0.0019244
#> 6   6      1   86 0.0347524 0.326062 0.512324 0.439863   6.90 0.0018952
#> 7   7      1   87 0.0356036 0.298270 0.527596 0.436170   6.71 0.0018279
#> 8   8      3   81 0.0163921 0.202899 0.869048 0.465753   8.45 0.0005939
#> 9   9      3   82 0.0190651 0.162218 0.772152 0.377049   5.71 0.0007047
#> 10 10      3   83 0.0151492 0.181586 1.028170 0.438356   8.69 0.0006587
#>     density    taxpc  region smsa   pctmin     wcon      wtuc     wtrd     wfir
#> 1  2.307159 25.69763 central   no 20.21870 206.4803  333.6209 182.3330 272.4492
#> 2  2.330254 24.87425 central   no 20.21870 212.7542  369.2964 189.5414 300.8788
#> 3  2.341801 26.45144 central   no 20.21870 219.7802 1394.8030 196.6395 309.9696
#> 4  2.346420 26.84235 central   no 20.21870 223.4238  398.8604 200.5629 350.0863
#> 5  2.364896 28.14034 central   no 20.21870 243.7562  358.7830 206.8827 383.0707
#> 6  2.385681 29.74098 central   no 20.21870 257.9139  369.5465 218.5165 409.8842
#> 7  2.422633 30.99368 central   no 20.21870 281.4259  408.7245 221.2701 453.1722
#> 8  0.976834 14.56088 central   no  7.91632 188.7683  292.6422 151.4234 202.4292
#> 9  0.992278 35.64073 central   no  7.91632 186.9658  345.7217 156.8826 225.0409
#> 10 1.003861 19.26188 central   no  7.91632 193.5983  604.9115 157.1295 248.1390
#>        wser   wmfg   wfed   wsta   wloc       mix   pctymle
#> 1  215.7335 229.12 409.37 236.24 231.47 0.0999179 0.0876968
#> 2  231.5767 240.33 419.70 253.88 236.79 0.1030491 0.0863767
#> 3  240.1568 269.70 438.85 250.36 248.58 0.0806787 0.0850909
#> 4  252.4477 281.74 459.17 261.93 264.38 0.0785035 0.0838333
#> 5  261.0861 298.88 490.43 281.44 288.58 0.0932486 0.0823065
#> 6  269.6129 322.65 478.67 286.91 306.70 0.0973228 0.0800806
#> 7  274.1775 334.54 477.58 292.09 311.91 0.0801688 0.0778710
#> 8  191.3742 210.75 381.72 247.38 213.17 0.0561224 0.0870046
#> 9  208.8190 217.77 386.42 374.07 219.18 0.0473118 0.0864722
#> 10 219.0847 236.64 382.65 268.90 223.06 0.0596206 0.0859426