Generate N=2 Data for Discrete and Bounded Continuous Examples

Generates arbitrary dataframes needed for discrete as well as continuous optimization problems when there are N individuals?

ffi_draw_n_alpha(
  fl_w_dollar = 100,
  it_w_units_avg = 5,
  fl_eh_max_inc_i = 5,
  fl_min_max_eh_inc_ratio = 1.5,
  fl_disc_cts_match_ratio = 1.1,
  fl_A_i_alpha_i1_ratio = 2,
  fl_A_i_relative_ratio = 3,
  it_N = 2,
  it_rand_seed = 123
)

Arguments

fl_w_dollar
float amount of aggregate resources in dollar for both individuals

it_w_units_avg
float number of goods for cts at max, approxi number for dis

fl_eh_max_inc_i
float maximum increase EH for one individual

fl_min_max_eh_inc_ratio
float ratio of maximum and minimum outcome increase

fl_disc_cts_match_ratio
float if 1, disc unit = it_w_units_avg

fl_A_i_alpha_i1_ratio
float A_i divide the alphi_il ratio, discrete set this

fl_A_i_relative_ratio
float Relative ratio of A between the two individuals

it_N
integer the number of individuals

it_rand_seed
integer random seed