Value at each W point along Queue, given optimal allocation

After solving for optimal allocating queue, at each Q point, which corresponds to different level of aggregate resources available, show the value given optimal choices this is an array of values, at each Q point. Potentially output for all Q. Or only show this for all queue points up to the actual resource limit, D_Wbin_il = 1.

ffp_opt_sodis_value(
  fl_rho,
  df_queue_il,
  bl_return_allQ_V = FALSE,
  bl_return_inner_V = FALSE,
  svr_id_i = "id_i",
  svr_D_il = "D_il",
  svr_inpalc = "Q_il",
  svr_D_Wbin_il = "D_Wbin_il",
  svr_A_il = "A_il",
  svr_alpha_il = "alpha_il",
  svr_beta_i = "beta_i",
  svr_V_cumu_l = "V_sum_l",
  svr_V_inner_Q_il = "V_inner_Q_il",
  svr_V_star_Q_il = "V_star_Q_il"
)

References

https://fanwangecon.github.io/PrjOptiAlloc/reference/ffp_opt_sodis_value.html https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_sodis_rkone_casch_allrw.html