Solves the linear allocation problem with Ai_i, alpha_i, and planner preferences.

This is the solution to the linear optimal allocation problem. Relative allocations, summed splined, inversion. solin, solution linear. relow, relative to the lowest allocation algoritm.

ffp_opt_solin_relow(
  df,
  svr_A_i,
  svr_alpha_i,
  svr_beta_i,
  fl_N_agg,
  fl_rho,
  svr_inpalc = "optiallocate",
  svr_expout = "optiexpoutcm"
)

Arguments

df
tibble data table including variables using svr names below each row is potentially an individual who will receive alternative allocations

svr_A_i
string name of the A_i variable, dot product of covariates and coefficients

svr_alpha_i
string name of the alpha_i variable, individual specific elasticity information

svr_beta_i
string name of the beta_i variable, relative preference weight for each child

fl_N_agg
float total resource avaible for allocation, if not specific, sum by svr_N_i

fl_rho
float preference for equality for the planner

svr_inpalc
string variable name for newly generated input optimal allocation, single word no dash

svr_expout
string variable name for newly generated expected outcome, single word no dash

Value

a dataframe that expands the df inputs with additional results.

a list with a dataframe and an array

df_opti - Dataframe with various statistcs related to optimal allocation, all intermediary stats

ar_opti_inpalc - Array where each element is an optimal choice for each individual

ar_opti_expout - Array where each element is the expected outcome given optimal choices for each i

References

https://fanwangecon.github.io/PrjOptiAlloc/reference/ffp_opt_solin_relow.html https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_solin_relow.html

Examples

data(df_opt_dtgch_cbem4)
df <- df_opt_dtgch_cbem4
svr_A_i <- 'A_lin'
svr_alpha_i <- 'alpha_lin'
svr_beta_i <- 'beta'
fl_N_agg <- 10000
fl_rho <- -30
ls_lin_solu <- ffp_opt_solin_relow(df, svr_A_i, svr_alpha_i, svr_beta_i, fl_N_agg, fl_rho)
df_opti <- ls_lin_solu$df_opti
ar_opti_inpalc <- ls_lin_solu$ar_opti_inpalc
ar_opti_expout <- ls_lin_solu$ar_opti_expout
summary(df_opti)

#>   S.country             vil.id         indi.id        svymthRound
#>  Length:1043        Min.   : 1.00   Min.   :   2.0   Min.   :24  
#>  Class :character   1st Qu.: 7.00   1st Qu.: 298.5   1st Qu.:24  
#>  Mode  :character   Median :12.00   Median : 627.0   Median :24  
#>                     Mean   :12.52   Mean   : 634.8   Mean   :24  
#>                     3rd Qu.:17.00   3rd Qu.: 939.5   3rd Qu.:24  
#>                     Max.   :33.00   Max.   :1349.0   Max.   :24  
#>      alpha           alpha_log              A             A_log      
#>  Min.   :0.01014   Min.   :0.006687   Min.   :73.70   Min.   :4.287  
#>  1st Qu.:0.01014   1st Qu.:0.006687   1st Qu.:77.73   1st Qu.:4.334  
#>  Median :0.06018   Median :0.012090   Median :78.45   Median :4.343  
#>  Mean   :0.04108   Mean   :0.010321   Mean   :78.46   Mean   :4.343  
#>  3rd Qu.:0.06018   3rd Qu.:0.012090   3rd Qu.:79.15   3rd Qu.:4.351  
#>  Max.   :0.06619   Max.   :0.013940   Max.   :84.61   Max.   :4.417  
#>       beta              rank_val        rank_idx      lowest_rank_A 
#>  Min.   :0.0009588   Min.   :101.0   Min.   :   1.0   Min.   :73.7  
#>  1st Qu.:0.0009588   1st Qu.:107.5   1st Qu.: 261.5   1st Qu.:73.7  
#>  Median :0.0009588   Median :109.7   Median : 522.0   Median :73.7  
#>  Mean   :0.0009588   Mean   :109.8   Mean   : 522.0   Mean   :73.7  
#>  3rd Qu.:0.0009588   3rd Qu.:112.4   3rd Qu.: 782.5   3rd Qu.:73.7  
#>  Max.   :0.0009588   Max.   :118.3   Max.   :1043.0   Max.   :73.7  
#>  rela_slope_to_lowest rela_intercept_to_lowest rela_slope_to_lowest_cumsum
#>  Min.   :0.9119       Min.   :-979.72          Min.   :   1.0             
#>  1st Qu.:1.0000       1st Qu.:-758.11          1st Qu.: 259.7             
#>  Median :1.0000       Median :-100.33          Median : 534.2             
#>  Mean   :2.6236       Mean   :-340.13          Mean   : 856.5             
#>  3rd Qu.:5.6008       3rd Qu.: -76.32          3rd Qu.:1349.4             
#>  Max.   :5.6008       Max.   :   0.00          Max.   :2736.4             
#>  rela_intercept_to_lowest_cumsum rela_slope_to_lowest_cumsum_invert
#>  Min.   :-354758                 Min.   :0.0003654                 
#>  1st Qu.:-143999                 1st Qu.:0.0007410                 
#>  Median : -41684                 Median :0.0018720                 
#>  Mean   : -90708                 Mean   :0.0069946                 
#>  3rd Qu.: -16952                 3rd Qu.:0.0038500                 
#>  Max.   :      0                 Max.   :1.0000000                 
#>  rela_intercept_to_lowest_cumsum_invert rela_x_intercept
#>  Min.   :  0.00                         Min.   :  0.00  
#>  1st Qu.: 65.27                         1st Qu.: 79.37  
#>  Median : 78.03                         Median :106.26  
#>  Mean   : 83.86                         Mean   :107.34  
#>  3rd Qu.:106.71                         3rd Qu.:138.90  
#>  Max.   :129.65                         Max.   :209.85  
#>  opti_lowest_spline_knots    tot_devi      allocate_lowest  optiallocate   
#>  Min.   :     0           Min.   :-10000   Min.   :96.12   Min.   : 0.000  
#>  1st Qu.:  3664           1st Qu.: -6336   1st Qu.:96.12   1st Qu.: 0.000  
#>  Median : 15077           Median :  5077   Median :96.12   Median : 0.000  
#>  Mean   : 25808           Mean   : 15808   Mean   :96.12   Mean   : 9.588  
#>  3rd Qu.: 43446           3rd Qu.: 33446   3rd Qu.:96.12   3rd Qu.:16.688  
#>  Max.   :219471           Max.   :209471   Max.   :96.12   Max.   :96.115  
#>  allocate_total   optiexpoutcm  
#>  Min.   :10000   Min.   :75.05  
#>  1st Qu.:10000   1st Qu.:78.35  
#>  Median :10000   Median :79.48  
#>  Mean   :10000   Mean   :79.04  
#>  3rd Qu.:10000   3rd Qu.:79.73  
#>  Max.   :10000   Max.   :84.61  

print(ar_opti_inpalc)

#>    [1]  0.07757637  0.08698173  0.09955769  0.22957298  0.37913904  0.40937073
#>    [7]  0.57345270  0.60795674  0.78881136  0.94574472  1.18772710  1.72333858
#>   [13]  1.72596536  1.83338817  1.86118932  1.99739447  2.36937909  2.21002823
#>   [19]  2.22658220  2.25865571  2.27164659  2.32435018  2.36268923  2.49921118
#>   [25]  2.75067459  2.75232017  3.03810958  2.91468088  3.08197835  3.14794474
#>   [31]  3.16655921  3.41367458  3.48902389  3.50471723  3.53843632  3.71599977
#>   [37]  3.73606358  3.92138680  3.98521701  4.01505622  4.18350867  4.63430784
#>   [43]  4.28194105  4.28213729  5.01812142  4.68493291 29.09408447  4.92274105
#>   [49]  5.58179423  5.16405903  5.84083063  6.02879352  6.19193048  5.67960670
#>   [55]  6.23528873  5.72368989  6.28782991  5.76214950  5.99376777  6.30171654
#>   [61]  6.30598890  6.37030971  6.42785194  6.53102484  6.61995374  6.79393166
#>   [67]  6.87878444  7.00501796  7.03350594  7.03660087  7.12020055  7.12292544
#>   [73]  7.33447990  8.06910174  7.53142033  7.59797544  7.79840328  7.87222756
#>   [79]  8.72175751  8.77789413  8.78121516  8.04649984  8.57290221  9.40242763
#>   [85]  8.65136889  8.69844889  8.69940765  9.63269419  8.83097040  8.91931059
#>   [91]  8.93645327  8.93839321  9.97909466  9.14270094 10.17345907 10.27500454
#>   [97]  9.41172195 10.45913145  9.57359206 10.62599646  9.77372554 10.76984561
#>  [103] 10.77784134 10.00708752 11.15908938 10.25827901 11.26136566 10.28143776
#>  [109] 10.34634728 10.38063265 10.39612975 10.40978502 10.42547836 11.50581362
#>  [115] 11.54641516 10.66689446 11.70847286 11.71295640 11.74801982 10.74157938
#>  [121] 10.81955546 10.83185951 11.90540282 66.88513801 10.96806466 12.06267789
#>  [127] 12.06384040 11.03044552 12.10646785 12.11030383 12.12062457 12.16006361
#>  [133] 12.19885508 11.16655255 11.19155312 11.20685398 11.22487974 11.29200112
#>  [139] 12.43871159 12.49006558 11.39198096 11.51229654 11.52246441 11.73848746
#>  [145] 11.78740929 12.94238712 11.96730516 11.96875451 13.14974575 13.17435582
#>  [151] 13.39919220 13.69861428 12.50746092 12.52838537 13.75199419 13.77298414
#>  [157] 12.56858868 12.64550789 13.91032384 12.71719599 12.81793835 12.81987829
#>  [163] 14.14880199 14.17576175 14.18343371 13.05159468 13.17714137 14.49468750
#>  [169] 13.23566479 14.53804576 13.26646275 14.62916256 14.64096959 13.39704431
#>  [175] 14.71044888 14.80496996 13.55756320 14.87796145 15.15296463 15.26644974
#>  [181] 14.00898629 15.66784900 15.67349504 15.70331945 15.76545057 15.83066218
#>  [187] 15.90003355 15.91399910 14.52502456 15.97345675 16.02075891 14.67886298
#>  [193] 14.69745501 14.70791724 16.14191597 16.18421965 16.28692763 16.34884290
#>  [199] 16.34927460 14.92443089 16.39892645 15.01849278 15.10383616 16.63003175
#>  [205] 16.64591528 16.72987508 16.75491686 17.01065691 17.07767858 15.58500034
#>  [211] 17.10804262 17.40149489 16.13426710 17.96975916 16.60157977 18.40184624
#>  [217] 18.42262033 16.82052396 18.46192676 18.52522039 18.68079332 17.03743008
#>  [223] 18.74654456 18.97265137 19.02305871 19.04042853 17.39506320 19.10115661
#>  [229] 19.16188469 19.16210054 19.16604445 19.23201154 19.29379420 19.31937561
#>  [235] 19.40301163 17.73682918 19.47591988 19.52651839 19.56231262 17.97959651
#>  [241] 19.72450292 19.76340233 19.82156486 19.95784998 19.98172925 19.98694359
#>  [247] 20.00995945 18.36272080 18.37492673 20.15977842 20.17885038 20.20410801
#>  [253] 20.39790811 20.40752274 18.64704267 20.46228100 20.48083801 20.67593322
#>  [259] 20.70554178 18.95587451 20.81754060 20.84322994 19.07957937 19.13189049
#>  [265] 21.03394954 19.18584720 19.21723387 21.09246054 21.11185627 21.26326945
#>  [271] 21.26421611 21.39688110 21.43267533 19.57334196 21.49437474 21.61307419
#>  [277] 19.76370005 21.80657519 22.06796128 22.18025920 22.27456442 22.30057753
#>  [283] 22.34309706 20.51698018 22.51944409 22.82140704 22.86168480 22.98176260
#>  [289] 23.08257727 23.12477302 23.15875719 23.16642914 23.20116879 23.27959048
#>  [295] 23.28631578 23.34758349 23.64209033 23.66553790 23.66564582 23.74917392
#>  [301] 23.91795692 23.96420450 23.99724201 24.00629232 24.12574725 24.21013875
#>  [307] 24.25009273 24.35334034 24.42762695 24.62293801 24.79436981 24.95899305
#>  [313] 22.77613371 25.36892767 25.42144417 25.47058107 25.73407632 25.87592423
#>  [319] 25.93249256 26.15562680 26.26238661 26.29818084 26.29861254 26.41966168
#>  [325] 26.51267179 26.79483197 26.81838746 26.97298906 26.99473449 27.03341805
#>  [331] 27.05759642 27.08647417 27.10416777 27.21284557 27.24769314 27.25377089
#>  [337] 27.33983985 27.40539992 27.76852265 27.79113149 27.81968547 28.74521058
#>  [343] 28.91547986 29.11965007 29.17038118 29.17527174 29.20307023 29.23878122
#>  [349] 29.27435959 29.31144893 29.59849967 29.86010161 30.05211633 30.13180845
#>  [355] 30.13267185 30.19702006 30.25871948 30.33927500 30.34981160 30.36474848
#>  [361] 30.45638023 30.72457487 30.75793615 31.02388902 31.07578264 31.30648091
#>  [367] 31.36623766 31.41005230 31.41472700 31.44457609 31.47769685 31.47983069
#>  [373] 31.50997888 31.63561948 31.72927714 31.97702146 31.99087908 32.01910928
#>  [379] 32.05490351 32.07005624 32.26547524 32.29381336 32.59301959 32.79578676
#>  [385] 32.96327465 32.98183166 33.03361735 33.08572682 33.17522473 33.23127811
#>  [391] 33.42796753 33.60207279 33.97907783 34.03672542 34.05622908 34.05717573
#>  [397] 34.10428672 34.52740673 34.79816691 34.91548800 34.98464351 35.09650973
#>  [403] 35.52135656 35.59210629 35.71058988 35.73052525 35.84452531 35.92243204
#>  [409] 36.05893301 36.38369597 36.43526581 37.89785242 37.99332015 39.04043403
#>  [415] 39.06931179 39.30692653 39.40485187 39.69360475 40.09095218 40.35169071
#>  [421] 40.53006366 41.48532993 41.61767115 41.72869864 41.84750601 41.99040850
#>  [427] 42.15203450 42.31230682 42.49184227 43.04700440 43.41823079 43.42536311
#>  [433] 43.52958206 43.75995655 44.31787539 44.73726735 45.33716609 45.54166008
#>  [439] 46.06216580 46.22786831 46.34367843 47.70258573 47.85389098 47.92104527
#>  [445] 48.12113896 49.11209154 50.86289140 51.31969648 51.85651745 52.00857818
#>  [451] 53.66688288 53.78077502 53.87518816 54.07957422 54.57568573 55.37570285
#>  [457] 55.69941122 56.74460711 59.43684030 59.75852277 60.61170394 62.79799515
#>  [463] 66.89981739 78.77746473 80.91304951 81.64666783 96.11543459  0.00000000
#>  [469]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [475]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [481]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [487]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [493]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [499]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [505]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [511]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [517]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [523]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [529]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [535]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [541]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [547]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [553]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [559]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [565]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [571]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [577]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [583]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [589]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [595]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [601]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [607]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [613]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [619]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [625]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [631]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [637]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [643]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [649]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [655]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [661]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [667]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [673]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [679]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [685]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [691]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [697]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [703]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [709]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [715]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [721]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [727]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [733]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [739]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [745]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [751]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [757]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [763]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [769]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [775]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [781]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [787]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [793]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [799]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [805]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [811]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [817]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [823]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [829]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [835]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [841]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [847]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [853]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [859]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [865]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [871]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [877]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [883]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [889]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [895]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [901]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [907]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [913]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [919]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [925]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [931]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [937]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [943]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [949]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [955]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [961]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [967]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [973]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [979]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [985]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [991]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#>  [997]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [1003]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [1009]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [1015]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [1021]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [1027]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [1033]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000
#> [1039]  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000

print(ar_opti_expout)

#>    [1] 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650
#>    [9] 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650
#>   [17] 79.48187 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650
#>   [25] 79.72650 79.72650 79.48187 79.72650 79.72650 79.72650 79.72650 79.72650
#>   [33] 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650
#>   [41] 79.72650 79.48187 79.72650 79.72650 79.48187 79.72650 75.04579 79.72650
#>   [49] 79.48187 79.72650 79.48187 79.48187 79.48187 79.72650 79.48187 79.72650
#>   [57] 79.48187 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650
#>   [65] 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650
#>   [73] 79.72650 79.48187 79.72650 79.72650 79.72650 79.72650 79.48187 79.48187
#>   [81] 79.48187 79.72650 79.72650 79.48187 79.72650 79.72650 79.72650 79.48187
#>   [89] 79.72650 79.72650 79.72650 79.72650 79.48187 79.72650 79.48187 79.48187
#>   [97] 79.72650 79.48187 79.72650 79.48187 79.72650 79.48187 79.48187 79.72650
#>  [105] 79.48187 79.72650 79.48187 79.72650 79.72650 79.72650 79.72650 79.72650
#>  [113] 79.72650 79.48187 79.48187 79.72650 79.48187 79.48187 79.48187 79.72650
#>  [121] 79.72650 79.72650 79.48187 75.04579 79.72650 79.48187 79.48187 79.72650
#>  [129] 79.48187 79.48187 79.48187 79.48187 79.48187 79.72650 79.72650 79.72650
#>  [137] 79.72650 79.72650 79.48187 79.48187 79.72650 79.72650 79.72650 79.72650
#>  [145] 79.72650 79.48187 79.72650 79.72650 79.48187 79.48187 79.48187 79.48187
#>  [153] 79.72650 79.72650 79.48187 79.48187 79.72650 79.72650 79.48187 79.72650
#>  [161] 79.72650 79.72650 79.48187 79.48187 79.48187 79.72650 79.72650 79.48187
#>  [169] 79.72650 79.48187 79.72650 79.48187 79.48187 79.72650 79.48187 79.48187
#>  [177] 79.72650 79.48187 79.48187 79.48187 79.72650 79.48187 79.48187 79.48187
#>  [185] 79.48187 79.48187 79.48187 79.48187 79.72650 79.48187 79.48187 79.72650
#>  [193] 79.72650 79.72650 79.48187 79.48187 79.48187 79.48187 79.48187 79.72650
#>  [201] 79.48187 79.72650 79.72650 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [209] 79.48187 79.72650 79.48187 79.48187 79.72650 79.48187 79.72650 79.48187
#>  [217] 79.48187 79.72650 79.48187 79.48187 79.48187 79.72650 79.48187 79.48187
#>  [225] 79.48187 79.48187 79.72650 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [233] 79.48187 79.48187 79.48187 79.72650 79.48187 79.48187 79.48187 79.72650
#>  [241] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.72650
#>  [249] 79.72650 79.48187 79.48187 79.48187 79.48187 79.48187 79.72650 79.48187
#>  [257] 79.48187 79.48187 79.48187 79.72650 79.48187 79.48187 79.72650 79.72650
#>  [265] 79.48187 79.72650 79.72650 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [273] 79.48187 79.72650 79.48187 79.48187 79.72650 79.48187 79.48187 79.48187
#>  [281] 79.48187 79.48187 79.48187 79.72650 79.48187 79.48187 79.48187 79.48187
#>  [289] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [297] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [305] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [313] 79.72650 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [321] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [329] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [337] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [345] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [353] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [361] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [369] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [377] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [385] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [393] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [401] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [409] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [417] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [425] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [433] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [441] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [449] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [457] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [465] 79.48187 79.48187 79.48187 79.73361 79.73949 79.74290 79.74462 79.75370
#>  [473] 75.09272 79.78232 79.78882 79.79048 79.80031 79.80101 79.80675 75.12223
#>  [481] 79.86010 79.86645 79.87024 79.88274 79.90824 79.66663 79.91848 79.95498
#>  [489] 79.96100 79.96759 79.97696 79.97840 79.99233 79.99614 80.00759 80.00903
#>  [497] 80.02731 80.05854 80.07188 75.37717 80.08700 80.11655 80.11942 80.12160
#>  [505] 80.12482 78.30534 80.17500 80.18398 80.18548 80.18778 80.18820 78.36651
#>  [513] 80.20961 80.23640 80.24788 80.27036 80.27731 80.27990 80.31003 75.61525
#>  [521] 80.33670 80.33879 78.49902 80.33999 80.36395 78.53825 75.67069 80.40841
#>  [529] 80.43347 78.59838 75.73533 80.47507 80.49621 78.67697 78.68392 78.68933
#>  [537] 78.70594 78.71795 80.56575 80.56593 80.58673 80.59173 80.59361 78.74854
#>  [545] 78.77002 75.88659 78.78695 78.80396 78.82172 78.83592 80.69424 78.85325
#>  [553] 78.86800 78.86823 78.87078 78.87194 78.87287 80.73063 78.88209 78.89190
#>  [561] 76.00335 78.89635 78.90078 80.75127 80.75767 78.90827 78.91993 78.92779
#>  [569] 78.93241 78.93804 76.05076 78.95007 78.95918 76.06796 78.96186 78.97775
#>  [577] 80.83032 78.98827 80.84773 79.00606 76.12665 76.12671 79.02993 80.88571
#>  [585] 79.03353 79.03756 79.03765 79.04114 79.05274 76.15721 80.90860 80.92110
#>  [593] 79.06816 80.92867 80.93852 80.94572 79.09631 76.21354 79.12616 79.13129
#>  [601] 79.13273 79.13894 76.25126 79.15316 79.15338 79.15546 81.01891 76.28828
#>  [609] 76.29115 76.29224 81.05269 79.19876 79.19981 79.20893 76.30844 76.32309
#>  [617] 76.32418 79.22959 79.23844 79.25681 79.26559 79.27642 81.15855 76.39433
#>  [625] 81.16084 81.16500 79.31479 79.31638 76.41246 79.32019 81.18115 76.41605
#>  [633] 79.32422 76.43025 79.33963 79.33995 79.34540 76.43804 79.35078 79.35601
#>  [641] 81.21669 79.37234 79.37470 79.37765 79.39466 79.40065 76.49277 79.44210
#>  [649] 79.44991 79.45076 76.54566 79.47582 79.47808 79.47947 79.48041 81.35327
#>  [657] 76.57810 79.49263 76.58325 79.49711 79.50418 76.60510 79.52303 79.52375
#>  [665] 79.52490 76.61658 76.61659 76.61918 79.53534 79.53994 81.40724 79.54352
#>  [673] 79.55294 79.56398 76.65557 79.57731 79.58435 79.59862 79.60146 76.69829
#>  [681] 79.61933 79.62682 79.62859 79.65780 76.75031 76.75675 79.69165 79.69735
#>  [689] 81.56994 79.70255 76.79124 79.72859 76.81346 79.73708 76.82148 76.82485
#>  [697] 79.75815 76.85751 76.85863 79.78821 76.86545 79.79678 79.80908 76.88868
#>  [705] 79.81400 79.81436 79.82265 76.89814 76.90105 79.84352 81.72991 79.86052
#>  [713] 76.93587 76.93914 79.86734 79.87958 79.88253 76.95930 81.76768 76.96770
#>  [721] 76.97398 76.97410 76.97503 76.98112 76.98962 76.99725 77.02021 79.95700
#>  [729] 77.03199 77.03324 77.03451 77.04278 77.06757 77.07086 77.07613 77.08902
#>  [737] 80.02808 77.11752 77.13914 81.95830 77.16132 77.17730 77.18297 77.18901
#>  [745] 77.19486 77.20135 77.20334 80.15564 80.15633 77.22010 77.22304 77.22735
#>  [753] 80.16541 77.23973 77.24234 77.25986 77.26350 77.27399 77.27682 80.21936
#>  [761] 77.28502 82.10624 77.30111 77.30315 80.24425 77.32263 77.32540 77.32702
#>  [769] 77.32958 77.34027 77.35112 77.36254 77.37350 77.38668 77.39062 80.35552
#>  [777] 77.42750 77.42873 80.37793 77.43854 77.45809 77.46988 82.31699 77.49317
#>  [785] 77.51610 80.46622 77.53989 77.54580 77.55284 77.55598 77.56532 80.52032
#>  [793] 77.57291 77.57799 77.59097 77.59099 77.59660 77.60119 77.60812 77.61631
#>  [801] 77.61664 77.62204 77.62369 80.57714 77.62612 77.62756 77.62891 77.63110
#>  [809] 77.63177 77.63213 77.63297 77.63792 77.63885 82.48301 77.64710 77.65687
#>  [817] 77.66538 77.66615 77.68510 77.68519 77.70306 77.70417 80.66141 80.66803
#>  [825] 77.72919 77.74582 77.74805 77.75001 77.75143 77.75550 77.75579 77.76104
#>  [833] 77.76381 80.72257 77.76772 77.77087 77.77210 80.73363 77.78097 77.79892
#>  [841] 77.80050 77.80788 77.80860 77.80883 77.81151 82.66567 77.81571 80.77672
#>  [849] 77.82360 77.82370 77.82972 77.83373 77.83778 77.83833 77.84043 77.84483
#>  [857] 77.84659 77.85771 77.86707 77.87311 77.87956 77.89290 77.89367 77.89820
#>  [865] 77.90775 77.91190 77.91305 77.91481 77.93637 77.93943 77.94175 77.94202
#>  [873] 77.94912 80.91568 77.95091 77.97456 77.97914 77.98546 77.99213 77.99844
#>  [881] 78.00086 78.00488 78.00809 78.01781 78.02450 81.00556 78.04462 78.06627
#>  [889] 78.07807 78.07829 78.07957 78.10010 78.10046 78.10260 78.10574 78.11050
#>  [897] 78.11672 78.11777 78.12668 78.12771 78.12803 78.12885 78.12898 78.13083
#>  [905] 78.13404 78.13666 78.13772 78.14546 78.14648 78.15364 78.15403 78.15454
#>  [913] 78.15990 78.17050 78.17562 78.17896 78.17959 78.19593 81.19769 78.22488
#>  [921] 78.22555 78.23050 78.23145 78.23374 78.24562 78.24722 78.24788 78.24811
#>  [929] 78.25378 78.25383 78.25650 78.26178 78.26257 78.26494 78.26499 78.26931
#>  [937] 78.27184 78.28061 78.28385 78.28530 78.28541 78.28645 78.28777 78.30397
#>  [945] 78.31157 78.32274 78.32474 78.32847 78.33612 78.34869 78.35063 78.35537
#>  [953] 78.36614 78.36829 78.37307 78.38329 78.39587 78.40152 78.40636 78.41266
#>  [961] 78.41340 78.41352 78.41607 78.42531 78.42745 78.42996 78.43560 78.43775
#>  [969] 78.44502 78.46064 78.46077 78.46454 78.47215 78.47367 78.48198 78.48729
#>  [977] 78.50209 78.50790 78.50797 78.50832 78.51844 78.53105 81.51984 78.54907
#>  [985] 78.54947 78.55185 78.55828 78.55875 78.56212 78.57038 78.58993 78.59732
#>  [993] 78.59735 78.60408 78.61014 78.61157 78.62410 78.63539 78.68103 78.68294
#> [1001] 78.68350 78.68774 78.69236 78.69330 78.70518 78.73668 78.74697 78.75868
#> [1009] 78.76318 78.78106 78.78464 81.78194 78.82434 78.83378 78.84038 78.87895
#> [1017] 78.88154 78.89056 78.89956 78.90591 81.93535 78.94894 78.95178 78.95991
#> [1025] 78.96625 78.97179 78.97390 78.97448 78.99004 79.03804 79.05826 79.06311
#> [1033] 79.08655 79.11331 82.14205 79.16941 79.17783 79.18255 79.20875 79.34690
#> [1041] 79.44826 79.52362 84.60875