Contains file processing components of ffp_snw_process_inputs function

ffp_snw_process_inputs solves allocation problems after first processing input files to generate MPCs. This file contains the initial parts of ffp_snw_process_inputs and only processes inputs files to generate MPCs so that MPC tables etc can be generated. This file does not proceed to do optimal allocation. Should be merged with ffp_snw_process_inputs so that code does not repeat.

ffp_snw_process_inputs_core(
 
    srt_simu_path = "C:/Users/fan/Documents/Dropbox (UH-ECON)/PrjNygaardSorensenWang/Output/",
  snm_simu_csv = "snwx_v_planner_moredense_a100zh266_e2m2.csv",
  df_plan_v_tilde_full = df_plan_v_tilde_full,
  fl_max_phaseout = 238000,
  it_bin_dollar_before_phaseout = 2500,
  it_bin_dollar_after_phaseout = 10000,
  fl_percheck_dollar = 100,
  fl_multiple = 58056,
  it_max_checks = 44,
  fl_tax_hh = 128580000,
  it_max_age = 64,
  it_min_age = 18,
  ar_ycut = NULL,
  ar_agecut = c(17, 29, 39, 49, 64),
  ar_svr_csv = c("age", "marital", "kids", "checks", "ymin", "mass", "survive",
    "vtilde", "ctilde"),
  ar_svr_groups = c("marital", "kids", "age_group", "ymin_group"),
  ar_svr_groups_stats = c("mass", "survive"),
  svr_checks = "checks",
  svr_v_value = "vtilde",
  svr_c_value = "ctilde",
  svr_mass = "mass",
  ar_rho = c(1),
  bl_threshold = FALSE,
  ls_stimulus_specs = list(st_biden_or_trump = "bidenchk", it_check_headorspouse = 12,
    it_check_perkids = 5),
  bl_given_firstcheck = FALSE,
  bl_non_inc_adjust = FALSE,
  bl_print = TRUE,
  bl_print_verbose = FALSE
)

Author

Fan Wang, http://fanwangecon.github.io