US 2008 Tax liability by household type given income for Nygaard, Sorernsen and Wang (2021)

We can study the effects of the 2008 Tax Rebate. The Tax rebate is a rebate based on how much tax was paid, so we need to know taxable income and tax liability. These differ by income, household marital status, and the count of children.

Given an array of pre-tax income values, we compute for from 0 to 4 kids and both married and unmarried taxable-income and tax-liability at all points along the income array.

Deductions are from 2008 income https://www.irs.gov/pub/irs-prior/f1040--2008.pdf. Tax brackets from 2008 are here https://www.irs.gov/pub/irs-prior/i1040tt--2008.pdf.

ffp_snw_tax_liability(
  ar_income = seq(0, 20000, length.out = 5),
  bl_verbose = FALSE
)

Arguments

ar_income

array of income points over which to evaluate taxable income and also tax liability by kids count and marital status.

Value

a list of 3-d arrays

mn_taxable_income - 2008 taxable income given income 3d array where 1st dimension is income, 2nd dimension is marital status, 3rd dimension is kids count

mn_tax_liability - 2008 taxable liability given income 3d array

References

https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_snw_stimulus_bush_2008.html

Author

Fan Wang, http://fanwangecon.github.io

Examples

ar_income <- c(1e4, 2e4, 4e4, 8e4, 1.6e5)
ls_taxable <- ffp_snw_tax_liability(ar_income)
mn_taxable_income <- ls_taxable$mn_taxable_income
mn_tax_liability <- ls_taxable$mn_tax_liability
print('mn_taxable_income')
#> [1] "mn_taxable_income"
print(mn_taxable_income)
#> , , kids=0
#> 
#>               married=0 married=1
#> income=10000       1050     -7900
#> income=20000      11050      2100
#> income=40000      31050     22100
#> income=80000      71050     62100
#> income=160000    151050    142100
#> 
#> , , kids=1
#> 
#>               married=0 married=1
#> income=10000      -5000    -11400
#> income=20000       5000     -1400
#> income=40000      25000     18600
#> income=80000      65000     58600
#> income=160000    145000    138600
#> 
#> , , kids=2
#> 
#>               married=0 married=1
#> income=10000      -8500    -14900
#> income=20000       1500     -4900
#> income=40000      21500     15100
#> income=80000      61500     55100
#> income=160000    141500    135100
#> 
#> , , kids=3
#> 
#>               married=0 married=1
#> income=10000     -12000    -18400
#> income=20000      -2000     -8400
#> income=40000      18000     11600
#> income=80000      58000     51600
#> income=160000    138000    131600
#> 
#> , , kids=4
#> 
#>               married=0 married=1
#> income=10000     -15500    -21900
#> income=20000      -5500    -11900
#> income=40000      14500      8100
#> income=80000      54500     48100
#> income=160000    134500    128100
#> 
print('mn_tax_liability')
#> [1] "mn_tax_liability"
print(mn_tax_liability)
#> , , kids=0
#> 
#>               married=0 married=1
#> income=10000     105.00       0.0
#> income=20000    1256.25     210.0
#> income=40000    4256.25    2512.5
#> income=80000   14106.25    8512.5
#> income=160000  36272.25   28532.0
#> 
#> , , kids=1
#> 
#>               married=0 married=1
#> income=10000        0.0       0.0
#> income=20000      500.0       0.0
#> income=40000     3177.5    1987.5
#> income=80000    11312.5    7987.5
#> income=160000   32283.0   27552.0
#> 
#> , , kids=2
#> 
#>               married=0 married=1
#> income=10000        0.0       0.0
#> income=20000      150.0       0.0
#> income=40000     2652.5    1510.0
#> income=80000    10437.5    7462.5
#> income=160000   31303.0   26572.0
#> 
#> , , kids=3
#> 
#>               married=0 married=1
#> income=10000        0.0       0.0
#> income=20000        0.0       0.0
#> income=40000     2127.5    1160.0
#> income=80000     9562.5    6937.5
#> income=160000   30323.0   25592.0
#> 
#> , , kids=4
#> 
#>               married=0 married=1
#> income=10000        0.0       0.0
#> income=20000        0.0       0.0
#> income=40000     1602.5     810.0
#> income=80000     8687.5    6412.5
#> income=160000   29343.0   24712.5
#>