Solves the linear allocation problem with init condi A, marg. prod alpha, and planner preferences. solin, solution linear. relow, relative to the lowest allocation algoritm.

This is the solution to the linear optimal allocation problem. Relative allocations, summed splined, inversion.

ffy_opt_solin_relow(df, svr_A_i, svr_alpha_i, svr_beta_i, fl_N_agg, fl_rho)

Arguments

df
tibble data table including variables using svr names below each row is potentially an individual who will receive alternative allocations

svr_A_i
string name of the A_i variable, dot product of covariates and coefficients

svr_alpha_i
string name of the alpha_i variable, individual specific elasticity information

svr_beta_i
string name of the beta_i variable, relative preference weight for each child

fl_N_agg
float total resource avaible for allocation, if not specific, sum by svr_N_i

fl_rho
float preference for equality for the planner

Value

a dataframe that expands the df inputs with additional results.

a list with a dataframe and an array

df_opti - Dataframe with various statistcs related to optimal allocation, all intermediary stats

ar_opti_inpalc - Array where each element is an optimal choice for each individual

ar_opti_expout - Array where each element is the expected outcome given optimal choices for each i

References

https://fanwangecon.github.io/REconTools/reference/ffy_opt_solin_relow.html https://fanwangecon.github.io/REconTools/articles/ffy_opt_solin_relow.html

Examples

data(df_opt_dtgch_cbem4)
df <- df_opt_dtgch_cbem4
svr_A_i <- 'A_lin'
svr_alpha_i <- 'alpha_lin'
svr_beta_i <- 'beta'
fl_N_agg <- 10000
fl_rho <- -30
ls_lin_solu <- ffy_opt_solin_relow(df, svr_A_i, svr_alpha_i, svr_beta_i, fl_N_agg, fl_rho)
df_opti <- ls_lin_solu$df_opti
ar_opti_inpalc <- ls_lin_solu$ar_opti_inpalc
ar_opti_expout <- ls_lin_solu$ar_opti_expout
summary(df_opti)

#>   S.country             vil.id         indi.id        svymthRound
#>  Length:1043        Min.   : 1.00   Min.   :   2.0   Min.   :24  
#>  Class :character   1st Qu.: 7.00   1st Qu.: 298.5   1st Qu.:24  
#>  Mode  :character   Median :12.00   Median : 627.0   Median :24  
#>                     Mean   :12.52   Mean   : 634.8   Mean   :24  
#>                     3rd Qu.:17.00   3rd Qu.: 939.5   3rd Qu.:24  
#>                     Max.   :33.00   Max.   :1349.0   Max.   :24  
#>                                                                  
#>      alpha           alpha_log              A             A_log      
#>  Min.   :0.01014   Min.   :0.006687   Min.   :73.70   Min.   :4.287  
#>  1st Qu.:0.01014   1st Qu.:0.006687   1st Qu.:77.73   1st Qu.:4.334  
#>  Median :0.06018   Median :0.012090   Median :78.45   Median :4.343  
#>  Mean   :0.04108   Mean   :0.010321   Mean   :78.46   Mean   :4.343  
#>  3rd Qu.:0.06018   3rd Qu.:0.012090   3rd Qu.:79.15   3rd Qu.:4.351  
#>  Max.   :0.06619   Max.   :0.013940   Max.   :84.61   Max.   :4.417  
#>                                                                      
#>       beta              rank_val        rank_idx      lowest_rank_A 
#>  Min.   :0.0009588   Min.   :101.0   Min.   :   1.0   Min.   :73.7  
#>  1st Qu.:0.0009588   1st Qu.:107.5   1st Qu.: 261.5   1st Qu.:73.7  
#>  Median :0.0009588   Median :109.7   Median : 522.0   Median :73.7  
#>  Mean   :0.0009588   Mean   :109.8   Mean   : 522.0   Mean   :73.7  
#>  3rd Qu.:0.0009588   3rd Qu.:112.4   3rd Qu.: 782.5   3rd Qu.:73.7  
#>  Max.   :0.0009588   Max.   :118.3   Max.   :1043.0   Max.   :73.7  
#>                                                                     
#>  rela_slope_to_lowest rela_intercept_to_lowest rela_slope_to_lowest_cumsum
#>  Min.   :0.9119       Min.   :-979.72          Min.   :   1.0             
#>  1st Qu.:1.0000       1st Qu.:-758.11          1st Qu.: 259.7             
#>  Median :1.0000       Median :-100.33          Median : 534.2             
#>  Mean   :2.6236       Mean   :-340.13          Mean   : 856.5             
#>  3rd Qu.:5.6008       3rd Qu.: -76.32          3rd Qu.:1349.4             
#>  Max.   :5.6008       Max.   :   0.00          Max.   :2736.4             
#>                                                                           
#>  rela_intercept_to_lowest_cumsum rela_slope_to_lowest_cumsum_invert
#>  Min.   :-354758                 Min.   :0.0003654                 
#>  1st Qu.:-143999                 1st Qu.:0.0007410                 
#>  Median : -41684                 Median :0.0018720                 
#>  Mean   : -90708                 Mean   :0.0069946                 
#>  3rd Qu.: -16952                 3rd Qu.:0.0038500                 
#>  Max.   :      0                 Max.   :1.0000000                 
#>                                                                    
#>  rela_intercept_to_lowest_cumsum_invert rela_x_intercept
#>  Min.   :  0.00                         Min.   :  0.00  
#>  1st Qu.: 65.27                         1st Qu.: 79.37  
#>  Median : 78.03                         Median :106.26  
#>  Mean   : 83.86                         Mean   :107.34  
#>  3rd Qu.:106.71                         3rd Qu.:138.90  
#>  Max.   :129.65                         Max.   :209.85  
#>                                                         
#>  opti_lowest_spline_knots    tot_devi      allocate_lowest opti_allocate    
#>  Min.   :     0           Min.   :-10000   Min.   :96.12   Min.   : 0.0776  
#>  1st Qu.:  3664           1st Qu.: -6336   1st Qu.:96.12   1st Qu.:11.1628  
#>  Median : 15077           Median :  5077   Median :96.12   Median :19.1319  
#>  Mean   : 25808           Mean   : 15808   Mean   :96.12   Mean   :21.4133  
#>  3rd Qu.: 43446           3rd Qu.: 33446   3rd Qu.:96.12   3rd Qu.:29.7293  
#>  Max.   :219471           Max.   :209471   Max.   :96.12   Max.   :96.1154  
#>                                                            NA's   :576      
#>  opti_allocate_total opti_exp_outcome
#>  Min.   :10000       Min.   :75.05   
#>  1st Qu.:10000       1st Qu.:79.48   
#>  Median :10000       Median :79.48   
#>  Mean   :10000       Mean   :79.54   
#>  3rd Qu.:10000       3rd Qu.:79.73   
#>  Max.   :10000       Max.   :79.73   
#>                      NA's   :576     

print(ar_opti_inpalc)

#>    [1]  0.07757637  0.08698173  0.09955769  0.22957298  0.37913904  0.40937073
#>    [7]  0.57345270  0.60795674  0.78881136  0.94574472  1.18772710  1.72333858
#>   [13]  1.72596536  1.83338817  1.86118932  1.99739447  2.36937909  2.21002823
#>   [19]  2.22658220  2.25865571  2.27164659  2.32435018  2.36268923  2.49921118
#>   [25]  2.75067459  2.75232017  3.03810958  2.91468088  3.08197835  3.14794474
#>   [31]  3.16655921  3.41367458  3.48902389  3.50471723  3.53843632  3.71599977
#>   [37]  3.73606358  3.92138680  3.98521701  4.01505622  4.18350867  4.63430784
#>   [43]  4.28194105  4.28213729  5.01812142  4.68493291 29.09408447  4.92274105
#>   [49]  5.58179423  5.16405903  5.84083063  6.02879352  6.19193048  5.67960670
#>   [55]  6.23528873  5.72368989  6.28782991  5.76214950  5.99376777  6.30171654
#>   [61]  6.30598890  6.37030971  6.42785194  6.53102484  6.61995374  6.79393166
#>   [67]  6.87878444  7.00501796  7.03350594  7.03660087  7.12020055  7.12292544
#>   [73]  7.33447990  8.06910174  7.53142033  7.59797544  7.79840328  7.87222756
#>   [79]  8.72175751  8.77789413  8.78121516  8.04649984  8.57290221  9.40242763
#>   [85]  8.65136889  8.69844889  8.69940765  9.63269419  8.83097040  8.91931059
#>   [91]  8.93645327  8.93839321  9.97909466  9.14270094 10.17345907 10.27500454
#>   [97]  9.41172195 10.45913145  9.57359206 10.62599646  9.77372554 10.76984561
#>  [103] 10.77784134 10.00708752 11.15908938 10.25827901 11.26136566 10.28143776
#>  [109] 10.34634728 10.38063265 10.39612975 10.40978502 10.42547836 11.50581362
#>  [115] 11.54641516 10.66689446 11.70847286 11.71295640 11.74801982 10.74157938
#>  [121] 10.81955546 10.83185951 11.90540282 66.88513801 10.96806466 12.06267789
#>  [127] 12.06384040 11.03044552 12.10646785 12.11030383 12.12062457 12.16006361
#>  [133] 12.19885508 11.16655255 11.19155312 11.20685398 11.22487974 11.29200112
#>  [139] 12.43871159 12.49006558 11.39198096 11.51229654 11.52246441 11.73848746
#>  [145] 11.78740929 12.94238712 11.96730516 11.96875451 13.14974575 13.17435582
#>  [151] 13.39919220 13.69861428 12.50746092 12.52838537 13.75199419 13.77298414
#>  [157] 12.56858868 12.64550789 13.91032384 12.71719599 12.81793835 12.81987829
#>  [163] 14.14880199 14.17576175 14.18343371 13.05159468 13.17714137 14.49468750
#>  [169] 13.23566479 14.53804576 13.26646275 14.62916256 14.64096959 13.39704431
#>  [175] 14.71044888 14.80496996 13.55756320 14.87796145 15.15296463 15.26644974
#>  [181] 14.00898629 15.66784900 15.67349504 15.70331945 15.76545057 15.83066218
#>  [187] 15.90003355 15.91399910 14.52502456 15.97345675 16.02075891 14.67886298
#>  [193] 14.69745501 14.70791724 16.14191597 16.18421965 16.28692763 16.34884290
#>  [199] 16.34927460 14.92443089 16.39892645 15.01849278 15.10383616 16.63003175
#>  [205] 16.64591528 16.72987508 16.75491686 17.01065691 17.07767858 15.58500034
#>  [211] 17.10804262 17.40149489 16.13426710 17.96975916 16.60157977 18.40184624
#>  [217] 18.42262033 16.82052396 18.46192676 18.52522039 18.68079332 17.03743008
#>  [223] 18.74654456 18.97265137 19.02305871 19.04042853 17.39506320 19.10115661
#>  [229] 19.16188469 19.16210054 19.16604445 19.23201154 19.29379420 19.31937561
#>  [235] 19.40301163 17.73682918 19.47591988 19.52651839 19.56231262 17.97959651
#>  [241] 19.72450292 19.76340233 19.82156486 19.95784998 19.98172925 19.98694359
#>  [247] 20.00995945 18.36272080 18.37492673 20.15977842 20.17885038 20.20410801
#>  [253] 20.39790811 20.40752274 18.64704267 20.46228100 20.48083801 20.67593322
#>  [259] 20.70554178 18.95587451 20.81754060 20.84322994 19.07957937 19.13189049
#>  [265] 21.03394954 19.18584720 19.21723387 21.09246054 21.11185627 21.26326945
#>  [271] 21.26421611 21.39688110 21.43267533 19.57334196 21.49437474 21.61307419
#>  [277] 19.76370005 21.80657519 22.06796128 22.18025920 22.27456442 22.30057753
#>  [283] 22.34309706 20.51698018 22.51944409 22.82140704 22.86168480 22.98176260
#>  [289] 23.08257727 23.12477302 23.15875719 23.16642914 23.20116879 23.27959048
#>  [295] 23.28631578 23.34758349 23.64209033 23.66553790 23.66564582 23.74917392
#>  [301] 23.91795692 23.96420450 23.99724201 24.00629232 24.12574725 24.21013875
#>  [307] 24.25009273 24.35334034 24.42762695 24.62293801 24.79436981 24.95899305
#>  [313] 22.77613371 25.36892767 25.42144417 25.47058107 25.73407632 25.87592423
#>  [319] 25.93249256 26.15562680 26.26238661 26.29818084 26.29861254 26.41966168
#>  [325] 26.51267179 26.79483197 26.81838746 26.97298906 26.99473449 27.03341805
#>  [331] 27.05759642 27.08647417 27.10416777 27.21284557 27.24769314 27.25377089
#>  [337] 27.33983985 27.40539992 27.76852265 27.79113149 27.81968547 28.74521058
#>  [343] 28.91547986 29.11965007 29.17038118 29.17527174 29.20307023 29.23878122
#>  [349] 29.27435959 29.31144893 29.59849967 29.86010161 30.05211633 30.13180845
#>  [355] 30.13267185 30.19702006 30.25871948 30.33927500 30.34981160 30.36474848
#>  [361] 30.45638023 30.72457487 30.75793615 31.02388902 31.07578264 31.30648091
#>  [367] 31.36623766 31.41005230 31.41472700 31.44457609 31.47769685 31.47983069
#>  [373] 31.50997888 31.63561948 31.72927714 31.97702146 31.99087908 32.01910928
#>  [379] 32.05490351 32.07005624 32.26547524 32.29381336 32.59301959 32.79578676
#>  [385] 32.96327465 32.98183166 33.03361735 33.08572682 33.17522473 33.23127811
#>  [391] 33.42796753 33.60207279 33.97907783 34.03672542 34.05622908 34.05717573
#>  [397] 34.10428672 34.52740673 34.79816691 34.91548800 34.98464351 35.09650973
#>  [403] 35.52135656 35.59210629 35.71058988 35.73052525 35.84452531 35.92243204
#>  [409] 36.05893301 36.38369597 36.43526581 37.89785242 37.99332015 39.04043403
#>  [415] 39.06931179 39.30692653 39.40485187 39.69360475 40.09095218 40.35169071
#>  [421] 40.53006366 41.48532993 41.61767115 41.72869864 41.84750601 41.99040850
#>  [427] 42.15203450 42.31230682 42.49184227 43.04700440 43.41823079 43.42536311
#>  [433] 43.52958206 43.75995655 44.31787539 44.73726735 45.33716609 45.54166008
#>  [439] 46.06216580 46.22786831 46.34367843 47.70258573 47.85389098 47.92104527
#>  [445] 48.12113896 49.11209154 50.86289140 51.31969648 51.85651745 52.00857818
#>  [451] 53.66688288 53.78077502 53.87518816 54.07957422 54.57568573 55.37570285
#>  [457] 55.69941122 56.74460711 59.43684030 59.75852277 60.61170394 62.79799515
#>  [463] 66.89981739 78.77746473 80.91304951 81.64666783 96.11543459          NA
#>  [469]          NA          NA          NA          NA          NA          NA
#>  [475]          NA          NA          NA          NA          NA          NA
#>  [481]          NA          NA          NA          NA          NA          NA
#>  [487]          NA          NA          NA          NA          NA          NA
#>  [493]          NA          NA          NA          NA          NA          NA
#>  [499]          NA          NA          NA          NA          NA          NA
#>  [505]          NA          NA          NA          NA          NA          NA
#>  [511]          NA          NA          NA          NA          NA          NA
#>  [517]          NA          NA          NA          NA          NA          NA
#>  [523]          NA          NA          NA          NA          NA          NA
#>  [529]          NA          NA          NA          NA          NA          NA
#>  [535]          NA          NA          NA          NA          NA          NA
#>  [541]          NA          NA          NA          NA          NA          NA
#>  [547]          NA          NA          NA          NA          NA          NA
#>  [553]          NA          NA          NA          NA          NA          NA
#>  [559]          NA          NA          NA          NA          NA          NA
#>  [565]          NA          NA          NA          NA          NA          NA
#>  [571]          NA          NA          NA          NA          NA          NA
#>  [577]          NA          NA          NA          NA          NA          NA
#>  [583]          NA          NA          NA          NA          NA          NA
#>  [589]          NA          NA          NA          NA          NA          NA
#>  [595]          NA          NA          NA          NA          NA          NA
#>  [601]          NA          NA          NA          NA          NA          NA
#>  [607]          NA          NA          NA          NA          NA          NA
#>  [613]          NA          NA          NA          NA          NA          NA
#>  [619]          NA          NA          NA          NA          NA          NA
#>  [625]          NA          NA          NA          NA          NA          NA
#>  [631]          NA          NA          NA          NA          NA          NA
#>  [637]          NA          NA          NA          NA          NA          NA
#>  [643]          NA          NA          NA          NA          NA          NA
#>  [649]          NA          NA          NA          NA          NA          NA
#>  [655]          NA          NA          NA          NA          NA          NA
#>  [661]          NA          NA          NA          NA          NA          NA
#>  [667]          NA          NA          NA          NA          NA          NA
#>  [673]          NA          NA          NA          NA          NA          NA
#>  [679]          NA          NA          NA          NA          NA          NA
#>  [685]          NA          NA          NA          NA          NA          NA
#>  [691]          NA          NA          NA          NA          NA          NA
#>  [697]          NA          NA          NA          NA          NA          NA
#>  [703]          NA          NA          NA          NA          NA          NA
#>  [709]          NA          NA          NA          NA          NA          NA
#>  [715]          NA          NA          NA          NA          NA          NA
#>  [721]          NA          NA          NA          NA          NA          NA
#>  [727]          NA          NA          NA          NA          NA          NA
#>  [733]          NA          NA          NA          NA          NA          NA
#>  [739]          NA          NA          NA          NA          NA          NA
#>  [745]          NA          NA          NA          NA          NA          NA
#>  [751]          NA          NA          NA          NA          NA          NA
#>  [757]          NA          NA          NA          NA          NA          NA
#>  [763]          NA          NA          NA          NA          NA          NA
#>  [769]          NA          NA          NA          NA          NA          NA
#>  [775]          NA          NA          NA          NA          NA          NA
#>  [781]          NA          NA          NA          NA          NA          NA
#>  [787]          NA          NA          NA          NA          NA          NA
#>  [793]          NA          NA          NA          NA          NA          NA
#>  [799]          NA          NA          NA          NA          NA          NA
#>  [805]          NA          NA          NA          NA          NA          NA
#>  [811]          NA          NA          NA          NA          NA          NA
#>  [817]          NA          NA          NA          NA          NA          NA
#>  [823]          NA          NA          NA          NA          NA          NA
#>  [829]          NA          NA          NA          NA          NA          NA
#>  [835]          NA          NA          NA          NA          NA          NA
#>  [841]          NA          NA          NA          NA          NA          NA
#>  [847]          NA          NA          NA          NA          NA          NA
#>  [853]          NA          NA          NA          NA          NA          NA
#>  [859]          NA          NA          NA          NA          NA          NA
#>  [865]          NA          NA          NA          NA          NA          NA
#>  [871]          NA          NA          NA          NA          NA          NA
#>  [877]          NA          NA          NA          NA          NA          NA
#>  [883]          NA          NA          NA          NA          NA          NA
#>  [889]          NA          NA          NA          NA          NA          NA
#>  [895]          NA          NA          NA          NA          NA          NA
#>  [901]          NA          NA          NA          NA          NA          NA
#>  [907]          NA          NA          NA          NA          NA          NA
#>  [913]          NA          NA          NA          NA          NA          NA
#>  [919]          NA          NA          NA          NA          NA          NA
#>  [925]          NA          NA          NA          NA          NA          NA
#>  [931]          NA          NA          NA          NA          NA          NA
#>  [937]          NA          NA          NA          NA          NA          NA
#>  [943]          NA          NA          NA          NA          NA          NA
#>  [949]          NA          NA          NA          NA          NA          NA
#>  [955]          NA          NA          NA          NA          NA          NA
#>  [961]          NA          NA          NA          NA          NA          NA
#>  [967]          NA          NA          NA          NA          NA          NA
#>  [973]          NA          NA          NA          NA          NA          NA
#>  [979]          NA          NA          NA          NA          NA          NA
#>  [985]          NA          NA          NA          NA          NA          NA
#>  [991]          NA          NA          NA          NA          NA          NA
#>  [997]          NA          NA          NA          NA          NA          NA
#> [1003]          NA          NA          NA          NA          NA          NA
#> [1009]          NA          NA          NA          NA          NA          NA
#> [1015]          NA          NA          NA          NA          NA          NA
#> [1021]          NA          NA          NA          NA          NA          NA
#> [1027]          NA          NA          NA          NA          NA          NA
#> [1033]          NA          NA          NA          NA          NA          NA
#> [1039]          NA          NA          NA          NA          NA

print(ar_opti_expout)

#>    [1] 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650
#>    [9] 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650
#>   [17] 79.48187 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650
#>   [25] 79.72650 79.72650 79.48187 79.72650 79.72650 79.72650 79.72650 79.72650
#>   [33] 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650
#>   [41] 79.72650 79.48187 79.72650 79.72650 79.48187 79.72650 75.04579 79.72650
#>   [49] 79.48187 79.72650 79.48187 79.48187 79.48187 79.72650 79.48187 79.72650
#>   [57] 79.48187 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650
#>   [65] 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650 79.72650
#>   [73] 79.72650 79.48187 79.72650 79.72650 79.72650 79.72650 79.48187 79.48187
#>   [81] 79.48187 79.72650 79.72650 79.48187 79.72650 79.72650 79.72650 79.48187
#>   [89] 79.72650 79.72650 79.72650 79.72650 79.48187 79.72650 79.48187 79.48187
#>   [97] 79.72650 79.48187 79.72650 79.48187 79.72650 79.48187 79.48187 79.72650
#>  [105] 79.48187 79.72650 79.48187 79.72650 79.72650 79.72650 79.72650 79.72650
#>  [113] 79.72650 79.48187 79.48187 79.72650 79.48187 79.48187 79.48187 79.72650
#>  [121] 79.72650 79.72650 79.48187 75.04579 79.72650 79.48187 79.48187 79.72650
#>  [129] 79.48187 79.48187 79.48187 79.48187 79.48187 79.72650 79.72650 79.72650
#>  [137] 79.72650 79.72650 79.48187 79.48187 79.72650 79.72650 79.72650 79.72650
#>  [145] 79.72650 79.48187 79.72650 79.72650 79.48187 79.48187 79.48187 79.48187
#>  [153] 79.72650 79.72650 79.48187 79.48187 79.72650 79.72650 79.48187 79.72650
#>  [161] 79.72650 79.72650 79.48187 79.48187 79.48187 79.72650 79.72650 79.48187
#>  [169] 79.72650 79.48187 79.72650 79.48187 79.48187 79.72650 79.48187 79.48187
#>  [177] 79.72650 79.48187 79.48187 79.48187 79.72650 79.48187 79.48187 79.48187
#>  [185] 79.48187 79.48187 79.48187 79.48187 79.72650 79.48187 79.48187 79.72650
#>  [193] 79.72650 79.72650 79.48187 79.48187 79.48187 79.48187 79.48187 79.72650
#>  [201] 79.48187 79.72650 79.72650 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [209] 79.48187 79.72650 79.48187 79.48187 79.72650 79.48187 79.72650 79.48187
#>  [217] 79.48187 79.72650 79.48187 79.48187 79.48187 79.72650 79.48187 79.48187
#>  [225] 79.48187 79.48187 79.72650 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [233] 79.48187 79.48187 79.48187 79.72650 79.48187 79.48187 79.48187 79.72650
#>  [241] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.72650
#>  [249] 79.72650 79.48187 79.48187 79.48187 79.48187 79.48187 79.72650 79.48187
#>  [257] 79.48187 79.48187 79.48187 79.72650 79.48187 79.48187 79.72650 79.72650
#>  [265] 79.48187 79.72650 79.72650 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [273] 79.48187 79.72650 79.48187 79.48187 79.72650 79.48187 79.48187 79.48187
#>  [281] 79.48187 79.48187 79.48187 79.72650 79.48187 79.48187 79.48187 79.48187
#>  [289] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [297] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [305] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [313] 79.72650 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [321] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [329] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [337] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [345] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [353] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [361] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [369] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [377] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [385] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [393] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [401] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [409] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [417] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [425] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [433] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [441] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [449] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [457] 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187 79.48187
#>  [465] 79.48187 79.48187 79.48187       NA       NA       NA       NA       NA
#>  [473]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [481]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [489]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [497]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [505]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [513]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [521]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [529]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [537]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [545]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [553]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [561]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [569]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [577]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [585]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [593]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [601]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [609]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [617]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [625]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [633]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [641]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [649]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [657]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [665]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [673]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [681]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [689]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [697]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [705]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [713]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [721]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [729]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [737]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [745]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [753]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [761]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [769]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [777]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [785]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [793]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [801]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [809]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [817]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [825]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [833]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [841]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [849]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [857]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [865]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [873]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [881]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [889]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [897]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [905]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [913]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [921]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [929]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [937]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [945]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [953]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [961]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [969]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [977]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [985]       NA       NA       NA       NA       NA       NA       NA       NA
#>  [993]       NA       NA       NA       NA       NA       NA       NA       NA
#> [1001]       NA       NA       NA       NA       NA       NA       NA       NA
#> [1009]       NA       NA       NA       NA       NA       NA       NA       NA
#> [1017]       NA       NA       NA       NA       NA       NA       NA       NA
#> [1025]       NA       NA       NA       NA       NA       NA       NA       NA
#> [1033]       NA       NA       NA       NA       NA       NA       NA       NA
#> [1041]       NA       NA       NA