Test Dataset Birth Weight

Dataset from Baystate Medical Center in Springfield, Massachusetts (Hosmer et. al. 1998)

data(df_opt_birthwt)

Format

csv

Source

https://onlinelibrary.wiley.com/doi/book/10.1002/9781118548387

References

HOSMER, D.W., HOSMER, T., LE CESSIE, S. and LEMESHOW, S. (1997), A COMPARISON OF GOODNESS‐OF‐FIT TESTS FOR THE LOGISTIC REGRESSION MODEL. Statist. Med., 16: 965-980. doi:10.1002/(SICI)1097-0258(19970515)16:9<965::AID-SIM509>3.0.CO;2-O

Examples

data(df_opt_birthwt)
head(df_opt_birthwt, 10)

#>    Index low age lwt race smoke ptl ht ui ftv  bwt
#> 1     85   0  19 182    2     0   0  0  1   0 2523
#> 2     86   0  33 155    3     0   0  0  0   3 2551
#> 3     87   0  20 105    1     1   0  0  0   1 2557
#> 4     88   0  21 108    1     1   0  0  1   2 2594
#> 5     89   0  18 107    1     1   0  0  1   0 2600
#> 6     91   0  21 124    3     0   0  0  0   0 2622
#> 7     92   0  22 118    1     0   0  0  0   1 2637
#> 8     93   0  17 103    3     0   0  0  0   1 2637
#> 9     94   0  29 123    1     1   0  0  0   1 2663
#> 10    95   0  26 113    1     1   0  0  0   0 2665