Test Dataset California Test Score Data: IL Frame, Generated Effect Each Additional Allocation

\(\text{ID}_i\): individual id \(\text{ID}_{il}\): unique id for individual/allocation id \(D^{\text{max}}_{i}\): maximimum discrete allocation each person \(D_{il}\): lth discrete allocation level for ith person, if fully redistributing, then this is equal to toal allocation count, if this is distribution additional teachers, this is equal to the additional allocation of teachers for each school. \(A_{il}\): A when considering the lth allocation for ith person, expected outcome without the lth allocation \(\alpha_{il}\): marginal expected effect of the lth allocation \(\beta_{i}\): i specific preference

data(df_opt_caschool_input_il)

Format

csv

Source

https://fmwww.bc.edu/ec-p/data/stockwatson/caschool.des https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_sodis_rkone_casch_allrw.html

References

Stock, James H. and Mark W. Watson (2003) Introduction to Econometrics, Addison-Wesley Educational Publishers, chapter 4–7.

Examples

data(df_opt_caschool_input_il)
head(df_opt_caschool_input_il, 30)
#> # A tibble: 30 x 7
#>     id_i id_il D_max_i  D_il  A_il alpha_il  beta_i
#>    <int> <int>   <dbl> <int> <dbl>    <dbl>   <dbl>
#>  1     1     1      11     1  691.    1.29  0.00238
#>  2     1     2      11     2  692.    1.09  0.00238
#>  3     1     3      11     3  693.    0.932 0.00238
#>  4     1     4      11     4  694.    0.808 0.00238
#>  5     1     5      11     5  695.    0.707 0.00238
#>  6     1     6      11     6  696.    0.624 0.00238
#>  7     1     7      11     7  696.    0.554 0.00238
#>  8     1     8      11     8  697.    0.496 0.00238
#>  9     1     9      11     9  697.    0.446 0.00238
#> 10     1    10      11    10  698.    0.404 0.00238
#> # ... with 20 more rows