Test Dataset California Test Score Data: PREP I Frame, Select Input and Estimation Dates

\(\text{ID}_i\): individual id \(D^{\text{max}}_{i}\): maximimum discrete allocation each person, in addition to existing levels \(D^{\text{o}}_{i}\): what I call observed allocation, just any comparison allocation, in addition to existing levels, this is the additional allocation (ignoring existing levels), this is the not the total allocation. If fully redistributing, this could be the observed level of allocation. \(\Omega_{i}\): This is the expected outcome when the input of interest is completely zero, \(\Omega_{i}\neq A_{i,l=0}\), because \(l=0\) does not mean input is zero, using \(\Omega_i\) rather than \(A_i\) to avoid confusion because \(A_i\) sounds like \(A_i = A_{i, l=0}\), but again, the problem does not start with allocation at zero, but each district already has some existing number of teachers. \(\theta_{i}\): coefficient in front of student teacher ratio \(\beta_{i}\): i specific preference \(\text{enrltot}_{i}\): enrollment per district/school, assume no within district variations \(\text{teachers}_{i}\): teacher per distrct/school \(\text{stravg}_{i}\): average of student teacher ratio district/school

data(df_opt_caschool_prep_i)

Format

csv

Source

https://fmwww.bc.edu/ec-p/data/stockwatson/caschool.des https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_sodis_rkone_casch_allrw.html

References

Stock, James H. and Mark W. Watson (2003) Introduction to Econometrics, Addison-Wesley Educational Publishers, chapter 4–7.

Examples

data(df_opt_caschool_prep_i)
head(df_opt_caschool_prep_i, 10)

#>    id_i D_max_i D_o_i  Omega_i    theta_i      beta_i enrltot teachers   stravg
#> 1     1       1     0 706.2218 -0.8699505 0.002380952     195       11 17.72727
#> 2     2       1     0 673.2570 -0.8699505 0.002380952     240       11 21.81818
#> 3     3       8     2 657.0958 -0.8699505 0.002380952    1550       83 18.67470
#> 4     4       1     0 656.3883 -0.8699505 0.002380952     243       14 17.35714
#> 5     5       7     1 655.5153 -0.8699505 0.002380952    1335       72 18.54167
#> 6     6       0     0 645.3442 -0.8699505 0.002380952     137        6 22.83333
#> 7     7       1     0 629.8211 -0.8699505 0.002380952     195       10 19.50000
#> 8     8       4     1 635.6090 -0.8699505 0.002380952     888       42 21.14286
#> 9     9       1     0 642.0002 -0.8699505 0.002380952     379       19 19.94737
#> 10   10      10     2 640.0966 -0.8699505 0.002380952    2247      108 20.80556
#>    distcod      county                        district grspan    avginc testscr
#> 1    75119     Alameda              Sunol Glen Unified  KK-08 22.690001  690.80
#> 2    61499       Butte            Manzanita Elementary  KK-08  9.824000  661.20
#> 3    61549       Butte     Thermalito Union Elementary  KK-08  8.978000  643.60
#> 4    61457       Butte Golden Feather Union Elementary  KK-08  8.978000  647.70
#> 5    61523       Butte        Palermo Union Elementary  KK-08  9.080333  640.85
#> 6    62042      Fresno         Burrel Union Elementary  KK-08 10.415000  605.55
#> 7    68536 San Joaquin           Holt Union Elementary  KK-08  6.577000  606.75
#> 8    63834        Kern             Vineland Elementary  KK-08  8.174000  609.00
#> 9    62331      Fresno        Orange Center Elementary  KK-08  7.385000  612.50
#> 10   67306  Sacramento     Del Paso Heights Elementary  KK-06 11.613333  612.65