Test Dataset Panel Height and Weight, Guatemala and Philippines, scrambled data, not actual.

Data from INCAP and CEBU, used in various projects, http://fanwangecon.github.io, note po stands for project optimal choice, to differentiate this from the files in REconTools

data(df_opt_dtgch_aorig)

Format

csv

Source

https://www.sciencedirect.com/science/article/pii/S1570677X16300107

References

"Early life height and weight production functions with endogenous energy and protein inputs", September 2016, 65-81, Economics & Human Biology, 22, Esteban Puentes, Fan Wang, Jere R. Behrman, Flavio Cunha, John Hoddinott, John A. Maluccio, Linda S. Adair, Judith Borja and Reynaldo Martorell

Examples

data(df_opt_dtgch_aorig)
var_wgt <- df_opt_dtgch_aorig$wgt
head(df_opt_dtgch_aorig, 10)
#> # A tibble: 10 x 15
#>    S.country vil.id indi.id sex   svymthRound momEdu wealthIdx   hgt   wgt  hgt0
#>    <chr>      <dbl>   <dbl> <chr>       <dbl>  <dbl>     <dbl> <dbl> <dbl> <dbl>
#>  1 Cebu           1       1 Male            0    5.3       6.3  45    985.  44.2
#>  2 Cebu           1       1 Male            2    5.3       6.3  54.6 3679.  44.2
#>  3 Cebu           1       1 Male            4    5.3       6.3  58.8 7954.  44.2
#>  4 Cebu           1       1 Male            6    5.3       6.3  65.6 6386.  44.2
#>  5 Cebu           1       1 Male            8    5.3       6.3  67.4 6039.  44.2
#>  6 Cebu           1       1 Male           10    5.3       6.3  70.4 8892.  44.2
#>  7 Cebu           1       1 Male           12    5.3       6.3  70.8 7325.  44.2
#>  8 Cebu           1       1 Male           14    5.3       6.3  NA     NA   44.2
#>  9 Cebu           1       1 Male           16    5.3       6.3  NA     NA   44.2
#> 10 Cebu           1       1 Male           18    5.3       6.3  NA     NA   44.2
#> # ... with 5 more variables: wgt0 <dbl>, prot <dbl>, cal <dbl>, p.A.prot <dbl>,
#> #   p.A.nProt <dbl>
var_wgt[1:10]
#>  [1]  985.4 3678.8 7954.4 6386.3 6038.7 8892.4 7325.1     NA     NA     NA