Test Dataset Job Training Effects Lalonde AER 1986

trt, a numeric vector identifying the study in which the subjects were enrolled (0 = Control, 1 = treated). Male module, 297 treated. 425 untreated. age, age (in years). educ, years of education. black, (0 = not black, 1 = black). hisp, (0 = not hispanic, 1 = hispanic). marr, (0 = not married, 1 = married). nodeg, (0 = completed high school, 1 = dropout). re74, real earnings in 1974. re75, real earnings in 1975. re78, real earnings in 1978.

data(df_opt_lalonde_training)

Format

csv

Source

https://www.jstor.org/stable/1806062?

References

Lalonde, R. (1986) "Evaluating the Econometric Evaluations of Training Programs with Experimental Data", American Economic Review, 604–620.

Examples

data(df_opt_lalonde_training)
head(df_opt_lalonde_training, 10)

#>     X trt age educ black hisp marr nodeg re74     re75      re78
#> 1   1   0  23   10     1    0    0     1    0    0.000     0.000
#> 2   2   0  26   12     0    0    0     0    0    0.000 12383.680
#> 3   3   0  22    9     1    0    0     1    0    0.000     0.000
#> 4   4   0  34    9     1    0    0     1   NA 4368.413 14051.160
#> 5   5   0  18    9     1    0    0     1    0    0.000 10740.080
#> 6   6   0  45   11     1    0    0     1    0    0.000 11796.470
#> 7   7   0  18    9     1    0    0     1    0    0.000  9227.052
#> 8   8   0  27   12     1    0    0     0   NA 2483.851  3552.268
#> 9   9   0  24    8     0    0    0     1    0    0.000 10569.270
#> 10 10   0  34   11     1    0    1     1    0    0.000  6040.335