Lalonde AER 1986 with Wage Regression Allocation Queue

appended onto the raw Lalonde binary targeting queue rank info based on different planner inequality aversion. includes min and max rank. Based on training logit regressin.

data(df_opt_lalonde_training_employ)

Format

csv

Source

https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_sobin_rkone_allfc_training_logit.html

References

Lalonde, R. (1986) "Evaluating the Econometric Evaluations of Training Programs with Experimental Data", American Economic Review, 604–620.

Examples

data(df_opt_lalonde_training_employ)
head(df_opt_lalonde_training_employ, 10)
#>    id A_i_employ alpha_i_employ beta_i_employ rank_min_employ rank_max_employ
#> 1   1  0.5906437     0.09506257   0.001385042              88               1
#> 2   2  0.8108942     0.04660746   0.001385042             605             587
#> 3   3  0.5975632     0.09429401   0.001385042              94               6
#> 4   4  0.6162286     0.07639905   0.001385042             294             146
#> 5   5  0.6356172     0.08947954   0.001385042             210              36
#> 6   6  0.6147834     0.07654263   0.001385042             286             133
#> 7   7  0.6356172     0.08947954   0.001385042             210              36
#> 8   8  0.6675923     0.07051941   0.001385042             463             290
#> 9   9  0.7926912     0.05022472   0.001385042             591             578
#> 10 10  0.6162294     0.07639897   0.001385042             295             147
#>    avg_rank_employ   trt age educ black hisp marr nodeg re74     re75      re78
#> 1         37.93333 ntran  23   10     1    0    0     1    0    0.000     0.000
#> 2        598.36667 ntran  26   12     0    0    0     0    0    0.000 12383.680
#> 3         45.16667 ntran  22    9     1    0    0     1    0    0.000     0.000
#> 4        213.30000 ntran  34    9     1    0    0     1   NA 4368.413 14051.160
#> 5        110.76667 ntran  18    9     1    0    0     1    0    0.000 10740.080
#> 6        202.96667 ntran  45   11     1    0    0     1    0    0.000 11796.470
#> 7        110.76667 ntran  18    9     1    0    0     1    0    0.000  9227.052
#> 8        376.90000 ntran  27   12     1    0    0     0   NA 2483.851  3552.268
#> 9        585.06667 ntran  24    8     0    0    0     1    0    0.000 10569.270
#> 10       214.30000 ntran  34   11     1    0    1     1    0    0.000  6040.335
#>    re75_zero emp78 emp75 race age_m2 age_m3     p_mpg  p_mpg_hp p_mpg_hp_bi0
#> 1          1     0     0    1      1      2 0.6783821 0.5906437    0.5906437
#> 2          1     1     0    0      2      2 0.8897568 0.8108942    0.8108942
#> 3          1     0     0    1      1      2 0.6883859 0.5975632    0.5975632
#> 4          0     1     1    1      2      3 0.6383402 0.6162286    0.6162286
#> 5          1     1     0    1      1      1 0.7042022 0.6356172    0.6356172
#> 6          1     1     0    1      2      3 0.5762456 0.6147834    0.6147834
#> 7          1     1     0    1      1      1 0.7042022 0.6356172    0.6356172
#> 8          0     1     1    1      2      3 0.7123512 0.6675923    0.6675923
#> 9          1     1     0    0      2      2 0.8749490 0.7926912    0.7926912
#> 10         1     1     0    1      2      3 0.6934077 0.6162294    0.6162294
#>    p_mpg_hp_bi1 rho_c1_rk_employ rho_c2_rk_employ rho_c3_rk_employ
#> 1     0.6857062                1                1                1
#> 2     0.8575016              605              605              605
#> 3     0.6918572                6                6                6
#> 4     0.6926277              294              294              294
#> 5     0.7250968               36               36               36
#> 6     0.6913261              286              286              274
#> 7     0.7250968               36               36               36
#> 8     0.7381117              463              463              463
#> 9     0.8429159              591              591              591
#> 10    0.6926284              295              295              295
#>    rho_c4_rk_employ rho_c5_rk_employ rho_c6_rk_employ rho_c7_rk_employ
#> 1                 1                1                1                1
#> 2               605              605              605              605
#> 3                 6                6                6                6
#> 4               294              282              282              273
#> 5                36               36               36               36
#> 6               274              274              273              260
#> 7                36               36               36               36
#> 8               463              463              454              454
#> 9               591              591              591              591
#> 10              295              283              283              274
#>    rho_c8_rk_employ rho_c9_rk_employ rho_c10_rk_employ rho_c11_rk_employ
#> 1                 1                1                 1                 1
#> 2               605              605               604               604
#> 3                 6                6                 6                 6
#> 4               268              268               268               253
#> 5                36               36                36                36
#> 6               260              260               256               241
#> 7                36               36                36                36
#> 8               454              450               447               447
#> 9               591              591               591               590
#> 10              269              269               269               254
#>    rho_c12_rk_employ rho_c13_rk_employ rho_c14_rk_employ rho_c15_rk_employ
#> 1                  1                 1                 1                 1
#> 2                604               604               604               604
#> 3                  6                 6                 6                 6
#> 4                246               233               217               212
#> 5                 36                53                90               103
#> 6                230               225               209               204
#> 7                 36                53                90               103
#> 8                444               435               416               397
#> 9                589               587               587               587
#> 10               247               234               218               213
#>    rho_c16_rk_employ rho_c17_rk_employ rho_c18_rk_employ rho_c19_rk_employ
#> 1                 12                45                64                68
#> 2                604               601               597               597
#> 3                 33                67                73                79
#> 4                211               207               199               177
#> 5                113               122               124               133
#> 6                203               198               191               169
#> 7                113               122               124               133
#> 8                375               349               330               314
#> 9                586               584               584               582
#> 10               212               208               200               178
#>    rho_c20_rk_employ rho_c21_rk_employ rho_c22_rk_employ rho_c23_rk_employ
#> 1                 75                78                81                87
#> 2                593               593               591               590
#> 3                 83                86                92                94
#> 4                155               155               147               147
#> 5                157               172               181               197
#> 6                147               147               139               139
#> 7                157               172               181               197
#> 8                307               297               293               293
#> 9                581               581               579               579
#> 10               156               156               148               148
#>    rho_c24_rk_employ rho_c25_rk_employ rho_c26_rk_employ rho_c27_rk_employ
#> 1                 87                87                87                88
#> 2                589               588               588               588
#> 3                 94                94                94                94
#> 4                147               146               146               146
#> 5                198               202               206               210
#> 6                138               138               136               133
#> 7                198               202               206               210
#> 8                292               292               292               290
#> 9                578               578               578               578
#> 10               148               147               147               147
#>    rho_c28_rk_employ rho_c29_rk_employ rho_c30_rk_employ
#> 1                 88                88                88
#> 2                588               588               587
#> 3                 94                94                94
#> 4                146               146               146
#> 5                210               210               210
#> 6                133               133               133
#> 7                210               210               210
#> 8                290               290               290
#> 9                578               578               578
#> 10               147               147               147