Lalonde AER 1986 with Wage Regression Allocation Queue Queue

appended onto the raw Lalonde binary targeting queue rank info based on different planner inequality aversion. includes min and max rank. Based on training logit regressin.

data(df_opt_lalonde_training_employ)

Format

csv

Source

https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_sobin_rkone_allfc_training_logit.html

References

Lalonde, R. (1986) "Evaluating the Econometric Evaluations of Training Programs with Experimental Data", American Economic Review, 604–620.

Examples

data(df_opt_lalonde_training_employ)
head(df_opt_lalonde_training_employ, 10)

#>    id  A_i_wage alpha_i_wage beta_i_wage rank_min_wage rank_max_wage
#> 1   1 0.6483406   0.07878288 0.001385042           180           187
#> 2   2 0.8577941   0.06358309 0.001385042           685           688
#> 3   3 0.6582975   0.07746136 0.001385042           235           246
#> 4   4 0.6062724   0.08370079 0.001385042            40            43
#> 5   5 0.6748927   0.07512792 0.001385042           406           423
#> 6   6 0.5438145   0.08893549 0.001385042             2             3
#> 7   7 0.6748927   0.07512792 0.001385042           406           423
#> 8   8 0.6744480   0.07519257 0.001385042           389           406
#> 9   9 0.8447118   0.06884295 0.001385042           622           640
#> 10 10 0.6630602   0.07680834 0.001385042           281           294
#>    avg_rank_wage   trt age educ black hisp marr nodeg re74     re75      re78
#> 1     182.666667 ntran  23   10     1    0    0     1    0    0.000     0.000
#> 2     687.000000 ntran  26   12     0    0    0     0    0    0.000 12383.680
#> 3     239.000000 ntran  22    9     1    0    0     1    0    0.000     0.000
#> 4      41.333333 ntran  34    9     1    0    0     1   NA 4368.413 14051.160
#> 5     412.333333 ntran  18    9     1    0    0     1    0    0.000 10740.080
#> 6       2.333333 ntran  45   11     1    0    0     1    0    0.000 11796.470
#> 7     412.333333 ntran  18    9     1    0    0     1    0    0.000  9227.052
#> 8     395.333333 ntran  27   12     1    0    0     0   NA 2483.851  3552.268
#> 9     634.000000 ntran  24    8     0    0    0     1    0    0.000 10569.270
#> 10    285.666667 ntran  34   11     1    0    1     1    0    0.000  6040.335
#>    emp78 emp75     p_mpg  p_mpg_hp p_mpg_hp_bi0 p_mpg_hp_bi1 rho_c1_rk
#> 1      0     0 0.6783821 0.6483406    0.6483406    0.7271235       180
#> 2      1     0 0.8897568 0.8577941    0.8577941    0.9213772       688
#> 3      0     0 0.6883859 0.6582975    0.6582975    0.7357589       235
#> 4      1     1 0.6383402 0.6062724    0.6062724    0.6899732        40
#> 5      1     0 0.7042022 0.6748927    0.6748927    0.7500207       406
#> 6      1     0 0.5762456 0.5438145    0.5438145    0.6327500         2
#> 7      1     0 0.7042022 0.6748927    0.6748927    0.7500207       406
#> 8      1     1 0.7123512 0.6744480    0.6744480    0.7496406       389
#> 9      1     0 0.8749490 0.8447118    0.8447118    0.9135548       640
#> 10     1     0 0.6934077 0.6630602    0.6630602    0.7398685       281
#>    rho_c2_rk rho_c3_rk
#> 1        181       187
#> 2        688       685
#> 3        236       246
#> 4         41        43
#> 5        408       423
#> 6          2         3
#> 7        408       423
#> 8        391       406
#> 9        640       622
#> 10       282       294