Lalonde AER 1986 with Wage Regression Allocation Queue

appended onto the raw Lalonde binary targeting queue rank info based on different planner inequality aversion. includes min and max rank. Based on wage regression.

data(df_opt_lalonde_training_wage)

Format

csv

Source

https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_sobin_rkone_allrw_training_wage.html

References

Lalonde, R. (1986) "Evaluating the Econometric Evaluations of Training Programs with Experimental Data", American Economic Review, 604–620.

Examples

data(df_opt_lalonde_training_wage)
head(df_opt_lalonde_training_wage, 10)

#>    id A_i_wage alpha_i_wage beta_i_wage rank_min_wage rank_max_wage
#> 1   1 4.705623    0.4537963 0.001385042           386           278
#> 2   2 6.842885    0.4537963 0.001385042           689           673
#> 3   3 4.373738    0.4537963 0.001385042           274           159
#> 4   4 4.129823    0.4537963 0.001385042           178            61
#> 5   5 4.192084    0.4537963 0.001385042           207            90
#> 6   6 5.057681    0.4537963 0.001385042           483           399
#> 7   7 4.192084    0.4537963 0.001385042           207            90
#> 8   8 5.096506    0.4537963 0.001385042           496           413
#> 9   9 5.445291    0.4537963 0.001385042           545           495
#> 10 10 5.450645    1.0708127 0.001385042           496            68
#>    avg_rank_wage trt age educ race marr nodeg      re78 rho_c1_rk_wage
#> 1       350.1000   0  23   10    1    0     1     0.000            386
#> 2       684.5000   0  26   12    0    0     0 12383.680            689
#> 3       231.5333   0  22    9    1    0     1     0.000            274
#> 4       132.4667   0  34    9    1    0     1 14051.160            178
#> 5       162.2667   0  18    9    1    0     1 10740.080            207
#> 6       456.1333   0  45   11    1    0     1 11796.470            483
#> 7       162.2667   0  18    9    1    0     1  9227.052            207
#> 8       470.1333   0  27   12    1    0     0  3552.268            496
#> 9       528.8000   0  24    8    0    0     1 10569.270            545
#> 10      216.7000   0  34   11    1    1     1  6040.335             68
#>    rho_c2_rk_wage rho_c3_rk_wage rho_c4_rk_wage rho_c5_rk_wage rho_c6_rk_wage
#> 1             386            386            386            386            386
#> 2             689            689            689            689            689
#> 3             274            274            274            274            274
#> 4             178            178            178            178            178
#> 5             207            207            207            207            207
#> 6             483            483            483            483            483
#> 7             207            207            207            207            207
#> 8             496            496            496            496            496
#> 9             545            545            545            545            545
#> 10             68             68             68             68             68
#>    rho_c7_rk_wage rho_c8_rk_wage rho_c9_rk_wage rho_c10_rk_wage rho_c11_rk_wage
#> 1             386            386            386             386             386
#> 2             689            689            689             689             689
#> 3             274            274            274             274             274
#> 4             178            178            178             178             178
#> 5             207            207            207             207             207
#> 6             483            483            483             483             483
#> 7             207            207            207             207             207
#> 8             496            496            496             496             496
#> 9             545            545            545             545             545
#> 10             68             68             68              68              68
#>    rho_c12_rk_wage rho_c13_rk_wage rho_c14_rk_wage rho_c15_rk_wage
#> 1              386             386             386             386
#> 2              689             689             689             689
#> 3              274             274             274             274
#> 4              178             178             178             178
#> 5              207             207             207             207
#> 6              483             483             483             483
#> 7              207             207             207             207
#> 8              496             496             496             496
#> 9              545             545             545             545
#> 10              68              68              68              68
#>    rho_c16_rk_wage rho_c17_rk_wage rho_c18_rk_wage rho_c19_rk_wage
#> 1              386             385             369             356
#> 2              689             689             689             689
#> 3              274             265             248             227
#> 4              178             162             146             112
#> 5              207             192             175             154
#> 6              483             483             473             460
#> 7              207             192             175             154
#> 8              496             496             487             473
#> 9              545             545             543             530
#> 10              68              68              85             199
#>    rho_c20_rk_wage rho_c21_rk_wage rho_c22_rk_wage rho_c23_rk_wage
#> 1              339             314             302             298
#> 2              688             686             682             675
#> 3              194             182             165             163
#> 4               86              68              64              61
#> 5              120              98              95              92
#> 6              451             437             432             416
#> 7              120              98              95              92
#> 8              466             456             447             437
#> 9              522             515             512             505
#> 10             316             366             445             481
#>    rho_c24_rk_wage rho_c25_rk_wage rho_c26_rk_wage rho_c27_rk_wage
#> 1              291             283             278             278
#> 2              674             674             673             673
#> 3              162             160             160             159
#> 4               61              61              61              61
#> 5               90              90              90              90
#> 6              407             401             400             399
#> 7               90              90              90              90
#> 8              423             416             414             414
#> 9              498             497             496             496
#> 10             485             491             494             495
#>    rho_c28_rk_wage rho_c29_rk_wage rho_c30_rk_wage
#> 1              278             278             278
#> 2              673             673             673
#> 3              159             159             159
#> 4               61              61              61
#> 5               90              90              90
#> 6              399             399             399
#> 7               90              90              90
#> 8              413             413             413
#> 9              495             495             495
#> 10             496             496             496