Lalonde AER 1986 with Wage Regression Allocation Queue

appended onto the raw Lalonde binary targeting queue rank info based on different planner inequality aversion. includes min and max rank. Based on wage regression.

data(df_opt_lalonde_training_wage)

Format

csv

Source

https://fanwangecon.github.io/PrjOptiAlloc/articles/ffv_opt_sobin_rkone_allrw_training_wage.html

References

Lalonde, R. (1986) "Evaluating the Econometric Evaluations of Training Programs with Experimental Data", American Economic Review, 604–620.

Examples

data(df_opt_lalonde_training_wage)
head(df_opt_lalonde_training_wage, 10)
#>    id A_i_wage alpha_i_wage beta_i_wage rank_min_wage rank_max_wage
#> 1   1 4.201030    0.3793939 0.001385042           247           135
#> 2   2 6.259851    0.3793939 0.001385042           648           622
#> 3   3 3.894398    0.3793939 0.001385042           185            68
#> 4   4 4.572335    0.3793939 0.001385042           327           222
#> 5   5 3.728447    0.3793939 0.001385042           159            42
#> 6   6 4.775996    0.3793939 0.001385042           381           286
#> 7   7 3.728447    0.3793939 0.001385042           159            42
#> 8   8 5.444597    0.3793939 0.001385042           549           485
#> 9   9 4.956678    0.3793939 0.001385042           428           339
#> 10 10 4.949491    1.0795020 0.001385042           338            28
#>    avg_rank_wage trt age educ race marr nodeg      re78 rho_c1_rk_wage
#> 1       206.0333   0  23   10    1    0     1     0.000            247
#> 2       640.5000   0  26   12    0    0     0 12383.680            648
#> 3       139.5667   0  22    9    1    0     1     0.000            185
#> 4       291.3333   0  34    9    1    0     1 14051.160            327
#> 5       112.0000   0  18    9    1    0     1 10740.080            159
#> 6       349.5000   0  45   11    1    0     1 11796.470            381
#> 7       112.0000   0  18    9    1    0     1  9227.052            159
#> 8       530.9000   0  27   12    1    0     0  3552.268            549
#> 9       399.9333   0  24    8    0    0     1 10569.270            428
#> 10      122.4000   0  34   11    1    1     1  6040.335             28
#>    rho_c2_rk_wage rho_c3_rk_wage rho_c4_rk_wage rho_c5_rk_wage rho_c6_rk_wage
#> 1             247            247            247            247            247
#> 2             648            648            648            648            648
#> 3             185            185            185            185            185
#> 4             327            327            327            327            327
#> 5             159            159            159            159            159
#> 6             381            381            381            381            381
#> 7             159            159            159            159            159
#> 8             549            549            549            549            549
#> 9             428            428            428            428            428
#> 10             28             28             28             28             28
#>    rho_c7_rk_wage rho_c8_rk_wage rho_c9_rk_wage rho_c10_rk_wage rho_c11_rk_wage
#> 1             247            247            247             247             247
#> 2             648            648            648             648             648
#> 3             185            185            185             185             185
#> 4             327            327            327             327             327
#> 5             159            159            159             159             159
#> 6             381            381            381             381             381
#> 7             159            159            159             159             159
#> 8             549            549            549             549             549
#> 9             428            428            428             428             428
#> 10             28             28             28              28              28
#>    rho_c12_rk_wage rho_c13_rk_wage rho_c14_rk_wage rho_c15_rk_wage
#> 1              247             247             247             247
#> 2              648             648             648             648
#> 3              185             185             185             185
#> 4              327             327             327             327
#> 5              159             159             159             159
#> 6              381             381             381             381
#> 7              159             159             159             159
#> 8              549             549             549             549
#> 9              428             428             428             428
#> 10              28              28              28              28
#>    rho_c16_rk_wage rho_c17_rk_wage rho_c18_rk_wage rho_c19_rk_wage
#> 1              247             244             226             205
#> 2              648             648             648             648
#> 3              185             175             156             113
#> 4              327             327             315             300
#> 5              159             145             119              72
#> 6              381             381             374             360
#> 7              159             145             119              72
#> 8              549             549             549             542
#> 9              428             428             425             408
#> 10              28              28              28              58
#>    rho_c20_rk_wage rho_c21_rk_wage rho_c22_rk_wage rho_c23_rk_wage
#> 1              173             152             142             138
#> 2              645             637             630             628
#> 3               86              76              73              70
#> 4              283             255             238             232
#> 5               53              47              44              42
#> 6              343             319             309             294
#> 7               53              47              44              42
#> 8              529             523             515             509
#> 9              399             386             362             356
#> 10             128             178             233             275
#>    rho_c24_rk_wage rho_c25_rk_wage rho_c26_rk_wage rho_c27_rk_wage
#> 1              138             136             135             135
#> 2              627             626             622             622
#> 3               70              68              68              68
#> 4              226             222             222             222
#> 5               42              42              42              42
#> 6              291             288             286             286
#> 7               42              42              42              42
#> 8              505             492             488             487
#> 9              349             341             340             339
#> 10             304             317             329             332
#>    rho_c28_rk_wage rho_c29_rk_wage rho_c30_rk_wage
#> 1              135             135             135
#> 2              622             622             622
#> 3               68              68              68
#> 4              222             222             222
#> 5               42              42              42
#> 6              286             286             286
#> 7               42              42              42
#> 8              485             485             485
#> 9              339             339             339
#> 10             338             338             338